### 2023-09-05 #5808 @editorialbot
[PRE REVIEW]: fff_segmenter: A signal segmentation script for acoustic FFF fabrication data in MATLAB
**Submitting author:** @thiago-glissoi (<a href="http://orcid.org/0000-0002-8860-2748">Thiago Glissoi Lopes</a>)
**Repository:** https://github.com/thiago-glissoi/FFF-Line-Segmentation
**Branch with paper.md** (empty if default branch):
**Version:** 1.0
