### 2023-09-05 #5808 @danielskatz
@thiago-glissoi - thanks for your submission.

Due to the relatively small amount of code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. You should hear back in a week or two.
