+++ fff_segmenter.m
@@ -194,14 +194,14 @@ adjustedContourLine1(1,3) = indexCountourTemp(1,3);
 adjustedContourLine1(1,2) = indexCountourTemp(1,3) - meanCountourGroup1;
 
 if indexCountourTemp(1,1) - meanCountourGroup1 < 1
-repoToContour1(1,1) = abs(indexCountourTemp(1,1) - meanCountourGroup1);
-repoToContour1(1,3) = adjustedContourLine1(1,2);
-repoToContour1(1,2) = repoToContour1(1,3)...
-    - repoToContour1(1,1);
+    repoToContour1(1,1) = abs(indexCountourTemp(1,1) - meanCountourGroup1);
+    repoToContour1(1,3) = adjustedContourLine1(1,2);
+    repoToContour1(1,2) = repoToContour1(1,3)...
+        - repoToContour1(1,1);
 else
-repoToContour1(1,1) = indexCountourTemp(1,1) - meanCountourGroup1;
-repoToContour1(1,3) = adjustedContourLine1(1,2);
-repoToContour1(1,2) = indexCountourTemp(1,2);
+    repoToContour1(1,1) = indexCountourTemp(1,1) - meanCountourGroup1;
+    repoToContour1(1,3) = adjustedContourLine1(1,2);
+    repoToContour1(1,2) = indexCountourTemp(1,2);
 end
 
 % reposition 2
@@ -240,7 +240,7 @@ contourRepositions = [repoToContour1; repoToContour2; repoToContour3];
 
 % 8° Obtaining first duration matrix for the internal pattern
 % There is no need to check for the problem of too many consecutive transition lines here
-% since there are very small line segments in the internal pattern. On the cobtrary, 
+% since there are very small line segments in the internal pattern. On the cobtrary,
 % this could cause some issues. So the result_problem value is set to false.
 resultProblem = false;
 
@@ -288,8 +288,8 @@ for i = 1:length (linesUdr9000)
     if  previousPeak < previousPeak2 % Change of behaviour from increasing to decreasing
         break;
     end
-    if  previousPeak > middleRasterDuration % Is larger than the middle raster duration, 
-    % found the end of the increasing behaviour
+    if  previousPeak > middleRasterDuration % Is larger than the middle raster duration,
+        % found the end of the increasing behaviour
         break;
     end
     if result == true % normal situation
@@ -401,6 +401,35 @@ if length(index_raster) < 55
 
 end
 
+% Verify external pattern integrity
+
+square1ExternalPattern = mean (indexContourAlt(1:4,1));
+square2ExternalPattern = mean (indexContourAlt(5:8,1));
+
+if abs(square2ExternalPattern - square1ExternalPattern) > 15e3
+    [contourRepositions, indexContourAlt] = adjustExternalPattern(contourRepositions, indexContourAlt);
+end
+
+%
+
+% Verify internal pattern integrity
+rasterAuxMatrix = index_raster;
+rasterAuxMatrix(1,4) = 0;
+
+for i = 2:length(index_raster(:,1))
+rasterAuxMatrix(i,4) = abs(index_raster(i,3) - index_raster(i-1,2));
+end
+
+maxRasterAux = max(rasterAuxMatrix(:,4));
+
+indexAdjust = find (rasterAuxMatrix(:,4) > 10e3);
+
+if maxRasterAux > 9e3
+    [index_raster, index_trans_raster] = adjustInternalPattern(index_raster, index_trans_raster, indexAdjust);
+end
+
+%
+
 
 signalReposition = Compos3r(sensorSignalNormalized,contourToRasterReposition(:,2:3))+...
     Compos3r(sensorSignalNormalized,contourRepositions(:,2:3));
[…]
