+++ fff_segmenter.m
@@ -4,9 +4,8 @@
 % which compose the internal pattern, and other movements of a signal obtained
 % from single-layer manufacturing by the Fused Filament Fabrication (FFF) process.
 %
-% % Version 2.5
-% Authorship: Thiago Glissoi Lopes - LADAPS - UNESP BAURU
-% Last edit: Thiago Glissoi Lopes - LADAPS - UNESP BAURU
+% MIT License
+% Copyright (c) 2023 Thiago Glissoi Lopes, Paulo Monteiro de Carvalho Monson, Paulo Roberto de Aguiar, Reinaldo Götz de Oliveira Junior, Pedro Oliveira Conceição Junior
 
 % CORE
 function fff_segmenter(sensorSignal, dirX, dirY, Fs)
@@ -35,6 +34,26 @@ if testSegmentChoicePoints ~= 1 && testSegmentChoiceSegments ~= 1
     segmentationChoice = 'points';
 end
 
+clear prompt
+clc;
+
+testSegmentChoice = strcmp(segmentationChoice,'points');
+
+if testSegmentChoice == true
+    disp ("Would you like to obtain the segmentation index (points)");
+    disp ("in number of samples or in seconds? ");
+    prompt = "Response number of samples/seconds [number of samples]: ";
+    xticksChoice = input(prompt,"s");
+    if isempty(xticksChoice)
+        xticksChoice = 'number of samples';
+    end
+    testxticksChoiceNumber = strcmp(xticksChoice,'number of samples');
+    testxticksChoiceSeconds = strcmp(xticksChoice,'seconds');
+    if testxticksChoiceNumber ~= 1 && testxticksChoiceSeconds ~= 1
+        xticksChoice = 'number of samples';
+    end
+
+end
 
 clear prompt
 clc;
@@ -96,6 +115,14 @@ disp ("Signal name: ");
 disp (signalIdentifier);
 disp ("Segmentation choice: ");
 disp (segmentationChoice);
+
+testSegmentChoice = strcmp(segmentationChoice,'points');
+
+if testSegmentChoice == true
+disp ("xticksChoice: ");
+disp (xticksChoice);
+end
+
 disp ("Graphical choice: ");
 disp (graphical);
 if graphical == 'Y'
@@ -388,6 +415,11 @@ contourToRasterReposition(1,2) = Duration4_v2(1,3);
 contourToRasterReposition(1,1) = contourToRasterReposition(1,2) -...
     contourToRasterReposition(1,3);
 
+contourToRasterRepositionAux = contourToRasterReposition;
+
+contourToRasterReposition(1,2) = contourToRasterRepositionAux(1,3);
+contourToRasterReposition(1,3) = contourToRasterRepositionAux(1,2);
+
 indexContourAlt = indexContour;
 
 for i = 1:1:12
@@ -417,7 +449,7 @@ rasterAuxMatrix = index_raster;
 rasterAuxMatrix(1,4) = 0;
 
 for i = 2:length(index_raster(:,1))
-rasterAuxMatrix(i,4) = abs(index_raster(i,3) - index_raster(i-1,2));
+    rasterAuxMatrix(i,4) = abs(index_raster(i,3) - index_raster(i-1,2));
 end
 
 maxRasterAux = max(rasterAuxMatrix(:,4));
@@ -443,15 +475,47 @@ signalContour = Compos3r(sensorSignalNormalized, indexContourAlt(:,2:3));
 
 testSegmentChoice = strcmp(segmentationChoice,'points');
 
+testxticksChoice = strcmp(xticksChoice,'seconds');
+
 if testSegmentChoice == true
     resultReposition = [contourRepositions; contourToRasterReposition];
     resultContour = indexContourAlt;
     resultRaster = index_raster;
[…]
