+++ readme.md
@@ -1,17 +1,18 @@
 # Definition
-The fff_segmenter algorithm is used to segment a acoustic signal obtained from a first layer 3D print into multiple acoustic blocks related to specific geometrical elements of the printed part, such as contour lines, raster lines, and transition between raster lines. The fff_segmenter algorithm uses the direction control signal of the X and Y axis stepper motors. In order to use the fff_segmenter algorithm, the acoustic signal must be collected synchronously with the X and Y direction control signals at a specified sampling frequency. We strongly encourage the user to read the ```Documentation``` pdf file in order to completely comprehend the fff_segmenter algorithm inner workings, and how it was built with a specific part geometry in consideration.
+The fff_segmenter algorithm is used to segment a acoustic signal obtained from a first layer 3D print into multiple acoustic blocks related to specific geometrical elements of the printed part, such as contour lines, raster lines, and transition between raster lines. The fff_segmenter algorithm uses the direction control signal of the X and Y axis stepper motors. In order to use the fff_segmenter algorithm, the acoustic signal must be collected synchronously with the X and Y direction control signals at a specified sampling frequency. The folder [data](data/) presents three datasets of signals collected from a first layer 3D print. 
+We strongly encourage the user to read the [Documentation](Documentation.pdf) pdf file in order to completely comprehend the fff_segmenter algorithm inner workings, and how it was built with a specific part geometry and sampling frequency in consideration.
 
 # Clone the repository
 git clone https://github.com/thiago-glissoi/FFF-Line-Segmentation.git
 
 # Example (run in Matlab)
-Let's load the data into the Matlab workspace
+Let's load the [Test1.mat](data/Test1.mat) dataset into the Matlab workspace
 
 ```Matlab
 load("Data/Test1.mat");
 ```
 
-Now, we run the fff_segmenter algorithm using the workspace variables of Test1.mat and a sampling frequency of 200 kHz
+Now, we run the fff_segmenter algorithm using the workspace variables of Test1.mat and the sampling frequency of 200 kHz
 
 ```Matlab
 fff_segmenter(Acoustic_signal,Dir_X,Dir_Y,200e3);
@@ -27,6 +28,13 @@ After that, the user will be prompted to choose the segmentation mode.
 
 ```Matlab
 Response points/segments [points]
+```
+
+After that, if the user opted for the points mode of segmentation, the user will be prompted to choose between obtaining the segmentation index in number of samples or in seconds.
+```Would you like to obtain the segmentation index (points) in number of samples or in seconds??```
+
+```Matlab
+Response number of samples/seconds [number of samples]
 ```
 
 After that, the user will be prompted to choose between obtaining a graphical visualization of the segmentation. 
@@ -50,21 +58,21 @@ Response Y/N [Y]
 
 After this last prompt, the fff_segmentation algorithm will run with the defined parameters. 
 
-For this example, i will use the default parameters as response to the prompts, and call the signal as ```'Test1'```, since I am using the Test1 data.
+In this example, default parameters are utilized in response to the prompts. The signal is designated as ```Test1``` to correspond with the Test1 dataset.
 
 The fff_segmentation algorithm will take between 30 seconds and 90 seconds to run.
-As a result of running the fff_segmentation algorithm with default parameters, I have obtained two new files in my current folder. The first, named 'points segmentation results 'Test1'.mat', holds the results of the segmentation in the index format. Figure 1 demonstrate the contents of the 'points segmentation results 'Test1'.mat' in the Matlab workspace.
[…]
