### 2023-10-23 9082e45
Adjust a minor error in the image
