### 2023-12-05 #5808 @thiago-glissoi
No problem with the delay. The authors appreciate your attention.

In order to help with this matter, I would like to suggest a few more researchers in the field.

Ismail Fidan
ifidan (at) tntech.edu

Sang Won Lee
sangwonl (at) skku.edu

Md. Hazrat Ali
md.ali (at) nu.edu.kz

Syuhei Kurokawa
kurobe- (at) mech.kyushu-u.ac.jp

Yan Wang
yan.wang (at) me.gatech.edu

Zhonghua Yu
yuzhh (at) zju.edu.cn

Hopefully these suggestions work well.

Best regards.

Thiago Glissoi Lopes
