### 2024-06-05 #6620 @thiago-glissoi
Hi @bmcfee

Thanks for the update. I noticed that @DARSakthi has started filling out the review checklist, but I haven't seen any reviewer-created issues on the [FFF-Line-Segmentation](https://github.com/thiago-glissoi/FFF-Line-Segmentation) repository yet.

Is there anything else we can do to gain momentum with the review?

Thanks!
