### 2024-06-06 thiago-glissoi/FFF-Line-Segmentation#2 @olivecha
Code comments and documentation 
While the existing documentation does a decent job at explaining what the code does, it is still hard to understand the code itself when reading the `fff_segmenter.m` file, and I don't think you could expect someone to be able to contribute to it without additional explanations from the author.

At the minimum, function documentation strings should be added and explain every argument, what is returned and what the function does, as described [here](https://www.mathworks.com/help/matlab/matlab_prog/add-help-for-your-program.html).

Also, the code would benefit from additional comments as it is a standalone script without formal API documentation.
