### 2024-06-06 thiago-glissoi/FFF-Line-Segmentation#4 @olivecha
Data structure for example usage
The example usage does not provide sufficient information for eventual users to use this code with their data. Please explain in more detail how `Data/Test...` were obtained and how could eventual users obtain similar data structures from their printing data.
