### 2024-06-06 thiago-glissoi/FFF-Line-Segmentation#5 @olivecha
User and `fff_segmenter` inputs
There is an inconsistency on how the input to `fff_segmenter` is processed versus the user input. The function call signature is API-like, requires MATLAB arrays to be defined beforehand and needs to be called from a MATLAB script or console, but it also requires user input.

I would suggest either:
- removing the arguments from the function call and adding a prompt that asks for the path of the data, which is then loaded from inside the function.
- Creating a second script meant to be edited by the user, where the data is loaded, and the function is called as-is.
