### 2024-06-06 thiago-glissoi/FFF-Line-Segmentation#6 @olivecha
Add at least one test
It would be important for users to validate that the script has the expected behaviour. You could simply add a testing script which performs the segmentation on a dataset supplied in the repository and compare the result against a saved previous result (regression test). This would catch unexpected changes in behaviours due to different MATLAB versions or operating systems.
