### 2024-06-06 #6620 @olivecha
As can be seen from the checklist, some work on the documentation should be done before publication, but the rest is in order. I don't think elaborate packaging should be expected from MATLAB projects.

The main issues requiring a certain amount of work which must be resolved before publication are:
- https://github.com/thiago-glissoi/FFF-Line-Segmentation/issues/2
- https://github.com/thiago-glissoi/FFF-Line-Segmentation/issues/4
- https://github.com/thiago-glissoi/FFF-Line-Segmentation/issues/5
- https://github.com/thiago-glissoi/FFF-Line-Segmentation/issues/6
