### 2024-06-06 thiago-glissoi/FFF-Line-Segmentation#2 @thiago-glissoi
Working on it. Thanks for the review.
