### 2024-06-10 b768898
Insert help comments to the main function and to the subfunctions
