+++ fff_segmenter.m
@@ -1,14 +1,60 @@
-% % READ ME
-% This function is used to automatically segment the manufacturing outline lines,
-% which make up the external pattern, the raster manufacturing lines,
-% which compose the internal pattern, and other movements of a signal obtained
-% from single-layer manufacturing by the Fused Filament Fabrication (FFF) process.
+function fff_segmenter(sensorSignal, dirX, dirY, Fs)
+%   fff_segmenter Is the main function of the FFF Line Segmentation script.
+%
+%% READ ME
+%   To verify the readmed, alonside a example run, check <a href="matlab: 
+%   web('https://github.com/thiago-glissoi/FFF-Line-Segmentation')">FFF-Line-Segmentation</a> for the FFF-Line-Segmentation script on Github.
 %
 % MIT License
 % Copyright (c) 2023 Thiago Glissoi Lopes, Paulo Monteiro de Carvalho Monson, Paulo Roberto de Aguiar, Reinaldo Götz de Oliveira Junior, Pedro Oliveira Conceição Junior
+%
+% The function performs the following steps:
+% 1. Obtains the output options based on user inputs.
+% 2. Performs the main data segmentation operations.
+% 3. Calls specific subfunctions to perform secondary tasks
+%    necessary to complete the segmentation operations.
+%
+%   For more information regarding the segmentation logic, see <a href="matlab: 
+%   web('https://github.com/thiago-glissoi/FFF-Line-Segmentation/blob/main/Documentation.pdf')">Documentation</a> for the FFF-Line-Segmentation script on Github.
+%
+% Inputs:
+% - sensorSignal (array): The acoustic signal obtained from the FFF process.
+% - dirX (array): The signal obtained from the X-axis step motor control.
+% - dirY (array): The signal obtained from the Y-axis step motor control.
+% - Fs (double): The sampling frequency of the signals.
+%
+% Outputs:
+% - The main outputs of the function are the segmentation results, which can be 
+%   the segmentation index (points), in seconds or in or the segments of signal (segments), based
+%   on the user's choice. These segmentation results are stored in the user current
+%   matlab workspace.
+% - Additionally, if the user opted for it, the segmentation results can be automatically 
+%   saved in .mat files in the user current matlab directory. Also, if the user opted for it,
+%   the function can generate a graphical visualization of the segmentation results and 
+%   automatically save it to the user current matlab directory.
+%
+% Called subfunctions (in alphabetical order):
+% <a href="matlab: disp('adjustExternalPattern: Adjust the external pattern segmentation results for a specific scenario. Please read the comments under the subfunction for further details.') ">adjustExternalPattern</a>
+% <a href="matlab: disp('adjustInternalPattern: Adjust the internal pattern segmentation results for a specific scenario. Please read the comments under the subfunction for further details.') ">adjustInternalPattern</a>
+% <a href="matlab: disp('adjust_internal: Adjust the internal pattern segmentation results for a specific scenario. Please read the comments under the subfunction for further details.') ">adjust_internal</a>
+% <a href="matlab: disp('Compos3r: Compose a signal based on the original acoustic signal and on specific segmentation indexes. Please read the comments under the subfunction for further details.') ">Compos3r</a>
+% <a href="matlab: disp('convUni: convUni Obtain a time (s) value based on a number of samples value, a reference signal, and on the sampling frequency. Please read the comments under the subfunction for further details.') ">convUni</a>
+% <a href="matlab: disp('detectAnom: Detect the anomaly verified in the Test1.mat dataset. Please read the comments under the subfunction for further details.') ">detectAnom</a>
[…]
