### 2024-06-10 5648939
Altering the code to include the values that are due to to the specific part geometry
