+++ fff_segmenter.m
@@ -54,6 +54,17 @@ function fff_segmenter(sensorSignal, dirX, dirY, Fs)
 % <a href="matlab: disp('save_files: Save the segmentation results in .mat files. Please read the comments under the subfunction for further details.') ">save_files</a>
 % <a href="matlab: disp('test_Minvariations: Test for the occurence of the variation problem observed for the Test1.mat data. Please read the comments under the subfunction for further details.') ">test_Minvariations</a>
 
+%% Attributing values that are due to the specific part geometry to necessary variables
+lowRefTransRaster = 8e3; % Lower Number of samples reference value for the transition raster
+uppRefTransRaster = 9e3; % Upper Number of samples reference value for the transition raster
+varDurRaster = 11e3; % Variation of the duration between raster lines
+adpDurTranRaster = 8.4e3; % Adopted duration for the transition raster
+numofRasterLines = 55; % Number of raster lines
+numofContourSides = 4; % Number of contour sides
+numofContourLoops = 3; % Number of contour loops
+numofContourLines = numofContourSides * numofContourLoops; % Number of contour lines
+durationReference = 150e3; % known duration for the last contour printing
+refminimum_sampleValue = 50e3; % Minimum number of samples reference for the test_Minvariations subfunction
 
 %% Input prompts
 clc;
@@ -211,7 +222,7 @@ dirXAdjustedNormalized = Normaliz3r(dirXAdjusted);
 
 % 3° Test for the occurence of the variation problem
 resultProblem = test_Minvariations(sensorSignalNormalized, dirXAdjustedNormalized,...
-    dirYAdjustedNormalized);
+    dirYAdjustedNormalized, lowRefTransRaster);
 % \ 3°
 
 clear dirXAdjusted dirX dirYAdjusted dirY sensorSignal
@@ -222,11 +233,10 @@ clear dirXAdjusted dirX dirYAdjusted dirY sensorSignal
 
 % 4° Obtain the duration vector
 Duration = obtainDuration(sensorSignalNormalized, dirXAdjustedNormalized,...
-    dirYAdjustedNormalized, resultProblem);
+    dirYAdjustedNormalized, resultProblem, refminimum_sampleValue);
 
 
 % 5° Find the separation point
-durationReference = 150e3; % known duration for the last contour printing
 
 for i = 1:length(Duration(:,1))
 
@@ -259,8 +269,8 @@ end
 % 7° Repositioning evaluation and contour lines correction
 
 % reposition 1
-meanCountourGroup1 = round(mean([indexCountourTemp(2,1) indexCountourTemp(3,1)...
-    indexCountourTemp(4,1)]),0);
+meanCountourGroup1 = round(mean([indexCountourTemp(numofContourSides-2,1) indexCountourTemp(numofContourSides-1,1)...
+    indexCountourTemp(numofContourSides,1)]),0);
 
 adjustedContourLine1(1,1) = meanCountourGroup1;
 adjustedContourLine1(1,3) = indexCountourTemp(1,3);
@@ -291,7 +301,7 @@ repoToContour2(1,2) = indexCountourTemp(5,2);
 
 % reposition 3
 meanContourGroup3 = round(mean([indexCountourTemp(10,1) indexCountourTemp(11,1)...
-    indexCountourTemp(12,1)]),0);
+    indexCountourTemp(numofContourLines,1)]),0);
 
 adjustedContourLine3(1,1) = meanContourGroup3;
 adjustedContourLine3(1,3) = indexCountourTemp(9,3);
@@ -321,18 +331,18 @@ resultProblem = false;
 Duration = obtainDuration(sensorSignalNormalized, ...
     dirXAdjustedNormalized,...
     dirYAdjustedNormalized, ...
-    resultProblem);
[…]
