### 2024-06-12 317f8f7
Alter the user input prompts to a dedicated GUI
