+++ APP_User_Inputs.m
@@ -1,0 +1,330 @@
+classdef APP_User_Inputs < matlab.apps.AppBase
+
+    % Properties that correspond to app components
+    properties (Access = public)
+        UIFigure                       matlab.ui.Figure
+        RunthesegmentationSwitchLabel  matlab.ui.control.Label
+        ObtainautomaticmatfilesSwitchLabel  matlab.ui.control.Label
+        RunthesegmentationSwitch       matlab.ui.control.Switch
+        ObtainautomaticmatfilesSwitch  matlab.ui.control.Switch
+        SavegraphicalvisualizationifobtainedLabel  matlab.ui.control.Label
+        ObtaingraphicalvisualizationSwitchLabel  matlab.ui.control.Label
+        SavegraphicalvisualizationifobtainedSwitch  matlab.ui.control.Switch
+        ObtaingraphicalvisualizationSwitch  matlab.ui.control.Switch
+        UnitforthePointssegmentationmodeButtonGroup  matlab.ui.container.ButtonGroup
+        SecondsButton                  matlab.ui.control.RadioButton
+        NumberofsamplesButton          matlab.ui.control.RadioButton
+        ChooseButton_3                 matlab.ui.control.RadioButton
+        SegmentationmodeButtonGroup    matlab.ui.container.ButtonGroup
+        SegmentsButton                 matlab.ui.control.RadioButton
+        PointsButton                   matlab.ui.control.RadioButton
+        ChooseButton                   matlab.ui.control.RadioButton
+        DirYidentificationEditField    matlab.ui.control.EditField
+        DirYidentificationEditFieldLabel  matlab.ui.control.Label
+        DirXidentificationEditField    matlab.ui.control.EditField
+        DirXidentificationEditFieldLabel  matlab.ui.control.Label
+        RawsignalidentificationEditField  matlab.ui.control.EditField
+        RawsignalidentificationEditFieldLabel  matlab.ui.control.Label
+        SamplingfrequencyEditField     matlab.ui.control.NumericEditField
+        SamplingfrequencyEditFieldLabel  matlab.ui.control.Label
+        SelectdataButton               matlab.ui.control.Button
+    end
+
+    % Callbacks that handle component events
+    methods (Access = private)
+
+        % Value changed function: ObtaingraphicalvisualizationSwitch
+        function ObtaingraphicalvisualizationSwitchValueChanged(app, event)
+            value = app.ObtaingraphicalvisualizationSwitch.Value;
+            if  strcmp(value,'Yes')
+                value2 = 'Y';
+            else
+                value2 = 'N';
+            end
+            assignin('base', 'graphical', value2);
+            
+        end
+
+        % Value changed function: 
+        % SavegraphicalvisualizationifobtainedSwitch
+        function SavegraphicalvisualizationifobtainedSwitchValueChanged(app, event)
+            value = app.SavegraphicalvisualizationifobtainedSwitch.Value;
+            if  strcmp(value,'Yes')
+                value3 = 'Y';
+            else
+                value3 = 'N';
+            end
+
+            assignin('base', 'figureChoice', value3);
+        end
+
+        % Value changed function: ObtainautomaticmatfilesSwitch
+        function ObtainautomaticmatfilesSwitchValueChanged(app, event)
+            value = app.ObtainautomaticmatfilesSwitch.Value;
+            if  strcmp(value,'Yes')
+                value4 = 'Y';
+            else
+                value4 = 'N';
+            end
+
+            assignin('base', 'saveChoice', value4);
+        end
+
+        % Value changed function: SamplingfrequencyEditField
[…]
