+++ fff_segmenter.m
@@ -1,4 +1,5 @@
-function fff_segmenter(sensorSignal, dirX, dirY, Fs)
+function fff_segmenter
+% function fff_segmenter(sensorSignal, dirX, dirY, Fs)
 %   fff_segmenter Is the main function of the FFF Line Segmentation script.
 %
 %% READ ME
@@ -44,7 +45,7 @@ function fff_segmenter(sensorSignal, dirX, dirY, Fs)
 % <a href="matlab: disp('gen_graph: Generate the graphical visualization of the segmentation results. Please read the comments under the subfunction for further details.') ">gen_graph</a>
 % <a href="matlab: disp('gen_signal_segments: Generate the signal segments based on the original signal and on specific segmentation indexes. Please read the comments under the subfunction for further details.') ">gen_signal_segments</a>
 % <a href="matlab: disp('generate_standard_fig: Plot the graphics with the segmentation results in a predefined format. Please read the comments under the subfunction for further details.') ">generate_standard_fig</a>
-% <a href="matlab: disp('isAnormal: Obtain the duration matrix for the internal pattern in an abnormal segmentation situation. Please read the comments under the subfunction for further details.') ">isAnormal</a>
+% <a href="matlab: disp('isAbnormal: Obtain the duration matrix for the internal pattern in an abnormal segmentation situation. Please read the comments under the subfunction for further details.') ">isAbnormal</a>
 % <a href="matlab: disp('isNormal: Obtain the duration matrix for the internal pattern in a normal segmentation situation. Please read the comments under the subfunction for further details.') ">isNormal</a>
 % <a href="matlab: disp('Normaliz3r: Normalize the amplitude of a signal between -1 and 1. Please read the comments under the subfunction for further details.') ">Normaliz3r</a>
 % <a href="matlab: disp('obtainDuration: Obtain the duration matrix based on the acoustic signal and on the step motor control signals. Please read the comments under the subfunction for further details.') ">obtainDuration</a>
@@ -54,70 +55,56 @@ function fff_segmenter(sensorSignal, dirX, dirY, Fs)
 % <a href="matlab: disp('save_files: Save the segmentation results in .mat files. Please read the comments under the subfunction for further details.') ">save_files</a>
 % <a href="matlab: disp('test_Minvariations: Test for the occurence of the variation problem observed for the Test1.mat data. Please read the comments under the subfunction for further details.') ">test_Minvariations</a>
 
-%% Attributing values that are due to the specific part geometry to necessary variables
-lowRefTransRaster = 8e3; % Lower Number of samples reference value for the transition raster
-uppRefTransRaster = 9e3; % Upper Number of samples reference value for the transition raster
-varDurRaster = 11e3; % Variation of the duration between raster lines
-adpDurTranRaster = 8.4e3; % Adopted duration for the transition raster
-numofRasterLines = 55; % Number of raster lines
-numofContourSides = 4; % Number of contour sides
-numofContourLoops = 3; % Number of contour loops
-numofContourLines = numofContourSides * numofContourLoops; % Number of contour lines
-durationReference = 150e3; % known duration for the last contour printing
-refminimum_sampleValue = 50e3; % Minimum number of samples reference for the test_Minvariations subfunction
 
-%% Input prompts
-clc;
-disp ("Provide the signal identification");
-prompt = "Signal name: ";
-signalIdentifier = input(prompt,"s");
[…]
