### 2024-06-16 b38b7f5
Major alterations to the documentation and other minor adjustments
