+++ fff_segmenter.m
@@ -55,39 +55,43 @@ function fff_segmenter
 % <a href="matlab: disp('save_files: Save the segmentation results in .mat files. Please read the comments under the subfunction for further details.') ">save_files</a>
 % <a href="matlab: disp('test_Minvariations: Test for the occurence of the variation problem observed for the Test1.mat data. Please read the comments under the subfunction for further details.') ">test_Minvariations</a>
 
+% app = APP_JOSS;
 
 %% APP User Inputs
 
-pause(0.1);
-
-% app = APP_JOSS;
 app = APP_User_Inputs;
 while strcmp(app.RunthesegmentationSwitch.Value,'Off') == 1
-    pause(0.1); 
+    pause(0.5); % In order to avoid missuse of the CPU
 end
 
 Data_path = evalin('base', 'Data_path');
 xticksChoice = evalin('base', 'xticksChoice');
-signalIdentifier = evalin('base', 'signalIdentifier');
 segmentationChoice = evalin('base','Segmentation_choice');
 saveChoice = evalin('base', 'saveChoice');
 graphical = evalin('base', 'graphical');
 Fs = evalin('base', 'Fs');
 figureChoice = evalin('base', 'figureChoice');
-
-load(Data_path);
-
+signalIdentifier = evalin('base', 'signalIdentifier');
 DirXidentification = evalin('base', 'DirXidentification');
 DirYidentification = evalin('base', 'DirYidentification');
-dirX = evalin('caller',DirXidentification);
+
+% Load variables from a .mat file into a temporary structure
+temp = load(Data_path); %#ok<*LOAD>
+
+% Assign each variable in the structure to the caller workspace
+fields = fieldnames(temp);
+for i = 1:length(fields)
+    assignin('caller', fields{i}, temp.(fields{i}));
+end
+clear temp fields i 
+
+dirX = evalin('base',DirXidentification);
 dirY = evalin('caller',DirYidentification);
 sensorSignal = evalin('caller', signalIdentifier);
 
 delete(app.UIFigure);
 
-%% Verify Input
-
-
+%% Verify Input integrity
 
 if isempty(signalIdentifier)
     signalIdentifier = 'segmentation results';
@@ -144,15 +148,18 @@ disp ("Wait just a few moments while we segment your FFF signal...");
 
 
 %% Attributing values that are due to the specific part geometry to necessary variables.
-% The Fs used for this values was 200e3;
-lowRefTransRaster = Fs/25; % Lower Number of samples reference value for the transition raster
-uppRefTransRaster = round(Fs/22.2222,0); % Upper Number of samples reference value for the transition raster
-varDurRaster = round(Fs/18.1818,0); % Variation of the duration between raster lines
-adpDurTranRaster = round(Fs/23.8095,0); % Adopted duration for the transition raster
+
+% Workpice specific variables
 numofRasterLines = 55; % Number of raster lines
 numofContourSides = 4; % Number of contour sides
 numofContourLoops = 3; % Number of contour loops
 numofContourLines = numofContourSides * numofContourLoops; % Number of contour lines
+
+% Variables for the segmentation logic
+lowRefTransRaster = Fs/25; % Lower Number of samples reference value for the transition raster
+uppRefTransRaster = round(Fs/22.2222,0); % Upper Number of samples reference value for the transition raster
+varDurRaster = round(Fs/18.1818,0); % Variation of the duration between raster lines
+adpDurTranRaster = round(Fs/23.8095,0); % Adopted duration for the transition raster
[…]
