### 2024-06-17 5d19e46
Alterations to the verifications of user input
