+++ fff_segmenter.m
@@ -64,85 +64,137 @@ while strcmp(app.RunthesegmentationSwitch.Value,'Off') == 1
     pause(0.5); % In order to avoid missuse of the CPU
 end
 
-Data_path = evalin('base', 'Data_path');
-xticksChoice = evalin('base', 'xticksChoice');
-segmentationChoice = evalin('base','Segmentation_choice');
-saveChoice = evalin('base', 'saveChoice');
-graphical = evalin('base', 'graphical');
+% Verify presence of identifications
+signalIdentifier_exists = evalin('base', ['exist(''', 'signalIdentifier', ''', ''var'')']);
+DirXidentification_exists = evalin('base', ['exist(''', 'DirXidentification', ''', ''var'')']);
+DirYidentification_exists = evalin('base', ['exist(''', 'DirYidentification', ''', ''var'')']);
+
+if signalIdentifier_exists == 1 && DirXidentification_exists == 1 &&...
+        DirYidentification_exists == 1
+else
+    disp ('ATTENTION!')
+        disp ('A necessary signal identification was not provided.');
+        disp ('Please, run the function again and provide the proper signal identifications.');
+        delete(app.UIFigure);
+        return;
+end
+
+% Verify presence of segmentation choice and xticks choice
+Segmentation_choice_exists = evalin('base', ['exist(''', 'Segmentation_choice', ''', ''var'')']);
+
+if Segmentation_choice_exists == 0
+    disp ('ATTENTION!')
+    disp ('The segmentation mode was not selected.');
+    disp ('The function will run with the default "Points" mode of segmentation');
+    segmentationChoice = 'Points';
+    xticksChoice_exists = evalin('base', ['exist(''', 'xticksChoice', ''', ''var'')']);
+    if xticksChoice_exists == 0
+        disp ('ATTENTION!')
+        disp ('The unit was not selected.');
+        disp ('The function will run with the default "Segments" unit');
+        xticksChoice = 'number of samples';
+    else
+        xticksChoice = evalin('base', 'xticksChoice');
+    end
+else
+    segmentationChoice = evalin('base','Segmentation_choice');
+
+    if strcmp(segmentationChoice,'Points')
+        xticksChoice_exists = evalin('base', ['exist(''', 'xticksChoice', ''', ''var'')']);
+        if xticksChoice_exists == 0
+            disp ('ATTENTION!')
+            disp ('The xticks mode was not selected.');
+            disp ('The function will run with the default "Segments" unit');
+            xticksChoice = 'number of samples';
+        else
+            xticksChoice = evalin('base', 'xticksChoice');
+        end
+
+    end
+end
+
+% Verify saveChoice
+saveChoice_exists = evalin('base', ['exist(''', 'saveChoice', ''', ''var'')']);
+if saveChoice_exists == 1
+    saveChoice = 'Y';
+else
+    saveChoice = 'N';
+end
+
+% Verify graphical representation
+graphical_exists = evalin('base', ['exist(''', 'graphical', ''', ''var'')']);
+if graphical_exists == 1
+    graphical = 'Y';
+    % Verify option to save figure
+    figureChoice_exists = evalin('base', ['exist(''', 'figureChoice', ''', ''var'')']);
+    if figureChoice_exists == 1
+        figureChoice = 'Y';
+    else
+        figureChoice = 'N';
+    end
+
+else
+    graphical = 'N';
+end
+
+% Verify Sampling frequency
+Fs_exists = evalin('base', ['exist(''', 'Fs', ''', ''var'')']);
+if Fs_exists == 1
 Fs = evalin('base', 'Fs');
[…]
