### 2024-06-17 c9c4232
Insert new auxiliary function and improve on user inputs verifications
