### 2024-06-24 #6620 @thiago-glissoi
@danielskatz  Yes, the authors are addressing the issues assigned by @olivecha .
Most of the alterations have already been made, and we are currently working on verifying the compatibility with Octave.
