+++ fff_segmenter.m
@@ -3,7 +3,7 @@ function fff_segmenter
 %   fff_segmenter Is the main function of the FFF Line Segmentation script.
 %
 %% READ ME
-%   To verify the readmed, alonside a example run, check <a href="matlab: 
+%   To verify the readmed, alonside a example run, check <a href="matlab:
 %   web('https://github.com/thiago-glissoi/FFF-Line-Segmentation')">FFF-Line-Segmentation</a> for the FFF-Line-Segmentation script on Github.
 %
 % MIT License
@@ -15,7 +15,7 @@ function fff_segmenter
 % 3. Calls specific subfunctions to perform secondary tasks
 %    necessary to complete the segmentation operations.
 %
-%   For more information regarding the segmentation logic, see <a href="matlab: 
+%   For more information regarding the segmentation logic, see <a href="matlab:
 %   web('https://github.com/thiago-glissoi/FFF-Line-Segmentation/blob/main/Documentation.pdf')">Documentation</a> for the FFF-Line-Segmentation script on Github.
 %
 % Inputs:
@@ -25,13 +25,13 @@ function fff_segmenter
 % - Fs (double): The sampling frequency of the signals.
 %
 % Outputs:
-% - The main outputs of the function are the segmentation results, which can be 
+% - The main outputs of the function are the segmentation results, which can be
 %   the segmentation index (points), in seconds or in or the segments of signal (segments), based
 %   on the user's choice. These segmentation results are stored in the user current
 %   matlab workspace.
-% - Additionally, if the user opted for it, the segmentation results can be automatically 
+% - Additionally, if the user opted for it, the segmentation results can be automatically
 %   saved in .mat files in the user current matlab directory. Also, if the user opted for it,
-%   the function can generate a graphical visualization of the segmentation results and 
+%   the function can generate a graphical visualization of the segmentation results and
 %   automatically save it to the user current matlab directory.
 %
 % Called subfunctions (in alphabetical order):
@@ -50,159 +50,225 @@ function fff_segmenter
 % <a href="matlab: disp('isNormal: Obtain the duration matrix for the internal pattern in a normal segmentation situation. Please read the comments under the subfunction for further details.') ">isNormal</a>
 % <a href="matlab: disp('Normaliz3r: Normalize the amplitude of a signal between -1 and 1. Please read the comments under the subfunction for further details.') ">Normaliz3r</a>
 % <a href="matlab: disp('obtainDuration: Obtain the duration matrix based on the acoustic signal and on the step motor control signals. Please read the comments under the subfunction for further details.') ">obtainDuration</a>
-% <a href="matlab: disp('obtainTable: Obtain a table with the segmentation results. Please read the comments under the subfunction for further details.') ">obtainTable</a>
+% <a href="matlab: disp('obtainOutput: Obtain a formated output with the segmentation results. Please read the comments under the subfunction for further details.') ">obtainOutput</a>
 % <a href="matlab: disp('obtain_time_vec: Obtain a time (s) vector based on the signal's number of samples length and on the sampling frequency. Please read the comments under the subfunction for further details.') ">obtain_time_vec</a>
 % <a href="matlab: disp('saveFig: Save the graphical visualization of the segmentation results in a predefined format and resolution. Please read the comments under the subfunction for further details.') ">saveFig</a>
[…]
