### 2024-06-25 bd89c1b
Correct minor problems with the segmentation logic
