+++ fff_segmenter.m
@@ -231,6 +231,7 @@ if exist('OCTAVE_VERSION', 'builtin') ~= 0
     disp (' ');
     graphical = input('Do you want to generate the graphical representation of the segmentation results? (Y/N): ', 's');
     if strcmp(graphical, 'Y') == 1
+        disp (' ');
         figureChoice = input('Do you want to save the graphical representation of the segmentation results? (Y/N): ', 's');
     end
     disp (' ');
@@ -238,6 +239,7 @@ if exist('OCTAVE_VERSION', 'builtin') ~= 0
     % Verify user input
     
     if strcmp(segmentationChoice,'Points') == 0 && strcmp(segmentationChoice,'Segments') == 0
+        disp (' ');
         disp ('ATTENTION!')
         disp ('The segmentation mode was not selected.');
         disp ('The function will run with the default "Points" mode of segmentation');
@@ -246,6 +248,7 @@ if exist('OCTAVE_VERSION', 'builtin') ~= 0
     end
     
     if strcmp(saveChoice,'Y') == 0 && strcmp(saveChoice,'N') == 0
+        disp (' ');
         disp ('ATTENTION!')
         disp ('The save choice was not selected.');
         disp ('The function will run with the default "N" mode of save');
@@ -253,20 +256,23 @@ if exist('OCTAVE_VERSION', 'builtin') ~= 0
     end
     
     if strcmp(graphical,'Y') == 0 && strcmp(graphical,'N') == 0
+        disp (' ');
         disp ('ATTENTION!')
         disp ('The graphical choice was not selected.');
         disp ('The function will run with the default "N" mode of graphical representation');
         graphical = 'N';
+        figureChoice = '';
     end
     
     if strcmp(figureChoice,'Y') == 0 && strcmp(figureChoice,'N') == 0
+        disp (' ');
         disp ('ATTENTION!')
         disp ('The figure choice was not selected.');
         disp ('The function will run with the default "N" mode of figure save');
         figureChoice = 'N';
     end
 end
-
+disp (' ');
 disp ("Wait just a few moments while we segment your FFF signal...");
 %#ok<*EXIST>
 
@@ -282,12 +288,13 @@ numofContourLines = numofContourSides * numofContourLoops; % Number of contour l
 lowRefTransRaster = Fs/25; % Lower Number of samples reference value for the transition raster
 
 if exist('OCTAVE_VERSION', 'builtin') == 0
-    uppRefTransRaster = round(Fs/22.2222,0); % Upper Number of samples reference value for the transition raster
+    % uppRefTransRaster = round(Fs/22.2222,0); % Upper Number of samples reference value for the transition raster
+    uppRefTransRaster = 9.3e3;
     varDurRaster = round(Fs/18.1818,0); % Variation of the duration between raster lines
     adpDurTranRaster = round(Fs/23.8095,0); % Adopted duration for the transition raster
     durationReference = round(Fs/1.33333,0); % known duration for the last contour printing
     refminimum_sampleValue = round(Fs/4,0); % Minimum number of samples reference for the test_Minvariations subfunction
-    middleRasterDuration = round(Fs/0.645161290322581,0); % Middle raster duration value
+    middleRasterDuration = round(Fs/0.645161290322581,1); % Middle raster duration value
 else
     uppRefTransRaster = round(Fs/22.2222); % Upper Number of samples reference value for the transition raster
     varDurRaster = round(Fs/18.1818); % Variation of the duration between raster lines
@@ -325,8 +332,6 @@ resultProblem = test_Minvariations(sensorSignalNormalized, dirXAdjustedNormalize
 clear dirXAdjusted dirX dirYAdjusted dirY sensorSignal
 
 %% External pattern segmentation
-
[…]
