### 2024-06-25 12235e6
Insert graphical input for Octave
