+++ fff_segmenter.m
@@ -94,7 +94,7 @@ if exist('OCTAVE_VERSION', 'builtin') == 0
             disp ('ATTENTION!')
             disp ('The unit was not selected.');
             disp ('The function will run with the default "Segments" unit');
-            xticksChoice = 'number of samples';
+            xticksChoice = 'Samples';
         else
             xticksChoice = evalin('base', 'xticksChoice');
         end
@@ -212,29 +212,32 @@ if exist('OCTAVE_VERSION', 'builtin') ~= 0
     pkg load signal
     pkg load control
     
-    dirX = evalin('base',input('Please enter the signal identification for the X axis: ', 's'));
     disp (' ');
-    dirY = evalin('base',input('Please enter the signal identification for the Y axis: ', 's'));
-    disp (' ');
-    sensorSignal = input('Please enter the signal identification for the acoustic signal: ', 's');
-    signalIdentifier = sensorSignal;
-    sensorSignal = evalin('base',sensorSignal);
-    disp (' ');
-    Fs = str2double(input('Please enter the sampling frequency of the signals: '));
-    disp (' ');
-    segmentationChoice = input('Do you want the segmentation results in points or segments? (Points/Segments): ', 's');
-    if strcmp(segmentationChoice,'Points') == 1
-        xticksChoice = input('Do you want the xticks in seconds or in number of samples? (Seconds/Samples): ', 's');
-    end
-    disp (' ');
-    saveChoice = input('Do you want to save the segmentation results? (Y/N): ', 's');
-    disp (' ');
-    graphical = input('Do you want to generate the graphical representation of the segmentation results? (Y/N): ', 's');
-    if strcmp(graphical, 'Y') == 1
-        disp (' ');
-        figureChoice = input('Do you want to save the graphical representation of the segmentation results? (Y/N): ', 's');
-    end
-    disp (' ');
+disp ('Load file with signals')
+[fname, fpath, fltidx] = uigetfile ();
+assignin( 'base', 'fpath', fpath);
+assignin( 'base', 'fname', fname);
+assignin( 'base', 'fltidx', fltidx);
+evalin( 'base', 'load ([fpath, fname]) ');
+prompt = {"Sensor signal identification", "X-axis signal identification",...
+"Y-axis signal identification", "Sampling frequency", "Segmentation type (Points/Segments)"
+"Segmentation unit (Samples/Seconds)", "Auto save files (Y/N)",...
+"Generate figure (Y/N)", "Save figure (Y/N)"};
+rowscols = [1,:; 1,:; 1,:; 1,:; 1,:; 1,:; 1,:; 1,:; 1,:];%#ok 
+defaults = {""; ""; ""; ""; "Points"; "Samples"; "N"; "N"; "N"};
+names = inputdlg(prompt, "FFF Line Segmentation - Insert values and choose between options",...
+ rowscols, defaults);
+
+signalIdentifier = names{1,1};
+sensorSignal = evalin( 'base' ,signalIdentifier);
+dirX = evalin( 'base', names {2,1}) ;
+dirY = evalin( 'base' ,names {3,1});
+Fs = str2double(names{4,1});
+segmentationChoice = names{5,1};
+xticksChoice = names{6,1};
+saveChoice = names {7,1};
+graphical = names{8,1};
+figureChoice = names{9,1};
     
     % Verify user input
     
@@ -288,20 +291,19 @@ numofContourLines = numofContourSides * numofContourLoops; % Number of contour l
 lowRefTransRaster = Fs/25; % Lower Number of samples reference value for the transition raster
 
 if exist('OCTAVE_VERSION', 'builtin') == 0
[…]
