+++ fff_segmenter.m
@@ -1,5 +1,4 @@
 function fff_segmenter
-% function fff_segmenter(sensorSignal, dirX, dirY, Fs)
 %   fff_segmenter Is the main function of the FFF Line Segmentation script.
 %
 %% READ ME
@@ -28,11 +27,11 @@ function fff_segmenter
 % - The main outputs of the function are the segmentation results, which can be
 %   the segmentation index (points), in seconds or in or the segments of signal (segments), based
 %   on the user's choice. These segmentation results are stored in the user current
-%   matlab workspace.
+%   matlab/octave workspace.
 % - Additionally, if the user opted for it, the segmentation results can be automatically
-%   saved in .mat files in the user current matlab directory. Also, if the user opted for it,
+%   saved in .mat files in the user current matlab/octave directory. Also, if the user opted for it,
 %   the function can generate a graphical visualization of the segmentation results and
-%   automatically save it to the user current matlab directory.
+%   automatically save it to the user current matlab/octave directory.
 %
 % Called subfunctions (in alphabetical order):
 % <a href="matlab: disp('adjustExternalPattern: Adjust the external pattern segmentation results for a specific scenario. Please read the comments under the subfunction for further details.') ">adjustExternalPattern</a>
@@ -59,17 +58,17 @@ function fff_segmenter
 if exist('OCTAVE_VERSION', 'builtin') == 0
     % Is not running in octave
     %% APP User Inputs
-    
+
     app = APP_User_Inputs;
     while strcmp(app.RunthesegmentationSwitch.Value,'Off') == 1
         pause(0.5); % In order to avoid missuse of the CPU
     end
-    
+
     % Verify presence of identifications
     signalIdentifier_exists = evalin('base', ['exist(''', 'signalIdentifier', ''', ''var'')']);
     DirXidentification_exists = evalin('base', ['exist(''', 'DirXidentification', ''', ''var'')']);
     DirYidentification_exists = evalin('base', ['exist(''', 'DirYidentification', ''', ''var'')']);
-    
+
     if signalIdentifier_exists == 1 && DirXidentification_exists == 1 &&...
             DirYidentification_exists == 1
     else
@@ -80,10 +79,10 @@ if exist('OCTAVE_VERSION', 'builtin') == 0
         cleanUpWS;
         return;
     end
-    
+
     % Verify presence of segmentation choice and xticks choice
     Segmentation_choice_exists = evalin('base', ['exist(''', 'Segmentation_choice', ''', ''var'')']);
-    
+
     if Segmentation_choice_exists == 0
         disp ('ATTENTION!')
         disp ('The segmentation mode was not selected.');
@@ -100,7 +99,7 @@ if exist('OCTAVE_VERSION', 'builtin') == 0
         end
     else
         segmentationChoice = evalin('base','Segmentation_choice');
-        
+
         if strcmp(segmentationChoice,'Points')
             xticksChoice_exists = evalin('base', ['exist(''', 'xticksChoice', ''', ''var'')']);
             if xticksChoice_exists == 0
@@ -111,10 +110,10 @@ if exist('OCTAVE_VERSION', 'builtin') == 0
             else
                 xticksChoice = evalin('base', 'xticksChoice');
             end
-            
+
         end
     end
-    
+
[…]
