+++ readme.md
@@ -1,78 +1,74 @@
 # Definition
-The fff_segmenter algorithm is used to segment a acoustic signal obtained from a first layer 3D print into multiple acoustic blocks related to specific geometrical elements of the printed part, such as contour lines, raster lines, and transition between raster lines. The fff_segmenter algorithm uses the direction control signal of the X and Y axis stepper motors. In order to use the fff_segmenter algorithm, the acoustic signal must be collected synchronously with the X and Y direction control signals at a specified sampling frequency. The folder [Data](Data/) presents three datasets of signals collected from a first layer 3D print. 
+The `fff_segmenter` algorithm is used to segment a acoustic signal obtained from a first layer 3D print into multiple acoustic blocks related to specific geometrical elements of the printed part, such as contour lines, raster lines, and transition between raster lines. The fff_segmenter algorithm uses the direction control signal of the X and Y axis stepper motors. In order to use the fff_segmenter algorithm, the acoustic signal must be collected synchronously with the X and Y direction control signals at a specified sampling frequency. The folder [Data](Data/) presents three datasets of signals collected from a first layer 3D print. 
 We strongly encourage the user to read the [Documentation](Documentation.pdf) pdf file in order to completely comprehend the fff_segmenter algorithm inner workings, and how it was built with a specific part geometry and sampling frequency in consideration.
 
 # Clone the repository
 git clone https://github.com/thiago-glissoi/FFF-Line-Segmentation.git
 
-# Example (run in Matlab)
-First load the [Test1.mat](Data/Test1.mat) dataset into the Matlab workspace
+# Dependencies
+</details>
+<details>
+  <summary><strong>Matlab dependencies</strong></summary>
+
+ -> [Signal Processing Toolbox](https://www.mathworks.com/products/signal.html)
+  </details>
+
+  </details>
+<details>
+  <summary><strong>Octave dependencies</strong></summary>
+
+ -> [Signal Package](https://octave.sourceforge.io/signal/)
+
+ -> [Control Package](https://octave.sourceforge.io/control/)
+  </details>
+
+# Example 
+Verify the example for Matlab or for Octave.
+
+</details>
+<details>
+  <summary><strong>Matlab example</strong></summary>
+
+Make sure that the listed [Dependencies](#Dependencies) are installed and loaded into Matlab.
+
+Run the `fff_segmenter` script by typing directly into the command window followed by the press of the Enter on the keyboard, or by clicking in the run button of the Matlab text editor.
 
 ```Matlab
-load("Data/Test1.mat");
+fff_segmenter
 ```
+After running the fff_segmenter algorithm, the user will be provided with the graphical interface displayed below.
 
-Now, run the fff_segmenter algorithm using the workspace variables of Test1.mat and the sampling frequency of 200 kHz
+![Figure 1 - Input interface](Example/Matlab%20Input%20GUI.png)
 
-```Matlab
-fff_segmenter(Acoustic_signal,Dir_X,Dir_Y,200e3);
-```
-After running the fff_segmenter algorithm, the user will be prompted to provide the signal identification.
-```Provide the signal identification```
-```Matlab 
-Signal name
-```
+For the purpose of this example, the `Select data` button will be used to load the `Test1.mat` file from the `Data` folder. The `Select data` button will open a file explorer window, where the user can navigate to the `Data` folder and select the `Test1.mat` file. The rest of the input fields will be filled with the values in regard to the `Test1.mat` dataset, with default values for the `Segmentation mode` and `Unit` segmentation options, and with all of the available outputs toggled to Yes.
 
-After that, the user will be prompted to choose the segmentation mode.
[…]
