### 2024-06-27 7a33357
Adjust Figure 2 image size
