### 2024-06-27 00400da
Update Figures 1 and 2 sizes
