### 2024-06-27 5986a93
Include specific figures to the Octave example run
