+++ readme.md
@@ -1,11 +1,13 @@
 # Definition
 The `fff_segmenter` algorithm is used to segment a acoustic signal obtained from a first layer 3D print into multiple acoustic blocks related to specific geometrical elements of the printed part, such as contour lines, raster lines, and transition between raster lines. The fff_segmenter algorithm uses the direction control signal of the X and Y axis stepper motors. In order to use the fff_segmenter algorithm, the acoustic signal must be collected synchronously with the X and Y direction control signals at a specified sampling frequency. The folder [Data](Data/) presents three datasets of signals collected from a first layer 3D print. 
-We strongly encourage the user to read the [Documentation](Documentation.pdf) pdf file in order to completely comprehend the fff_segmenter algorithm inner workings, and how it was built with a specific part geometry and sampling frequency in consideration.
+We `strongly encourage` the user to read the [Documentation](Documentation.pdf) pdf file in order to completely comprehend the fff_segmenter algorithm inner workings, and how it was built with a specific part geometry and sampling frequency in consideration.
 
 # Clone the repository
 git clone https://github.com/thiago-glissoi/FFF-Line-Segmentation.git
 
 # Dependencies
+Before using the fff_segmenter script, verify if the following dependencies are installed and available in your Matlab or Octave environment.
+
 </details>
 <details>
   <summary><strong>Matlab dependencies</strong></summary>
@@ -38,40 +40,75 @@ fff_segmenter
 ```
 After running the fff_segmenter algorithm, the user will be provided with the graphical interface displayed below.
 
-![]()
+<img src="Example/Matlab%20Input%20GUI.png" alt="Figure 1 - Input interface" width="420">
 
-<img src="Example/Matlab%20Input%20GUI.png" alt="Figure 1 - Input interface" width="300">
+For the purpose of this example, the `Select data` button will be used to load the `Test1.mat` file from the `Data` folder. The `Select data` button will open a file explorer window, where the user can navigate to the `Data` folder and select the `Test1.mat` file. The rest of the input fields will be filled with the values in regard to the `Test1.mat` dataset, with default values for the `Segmentation mode` and `Unit` segmentation options, and with all of the available outputs toggled to `Yes`.
 
-For the purpose of this example, the `Select data` button will be used to load the `Test1.mat` file from the `Data` folder. The `Select data` button will open a file explorer window, where the user can navigate to the `Data` folder and select the `Test1.mat` file. The rest of the input fields will be filled with the values in regard to the `Test1.mat` dataset, with default values for the `Segmentation mode` and `Unit` segmentation options, and with all of the available outputs toggled to Yes.
-
-<img src="Example/Matlab%20Input%20GUI%20filled.png" alt="Figure 2 - Input interface filled with values" width="300">
+<img src="Example/Matlab%20Input%20GUI%20filled.png" alt="Figure 2 - Input interface filled with values" width="420">
 
 
 When the user toggle to `On` the `Run the segmentation` button, the graphical input interface will be closed and the fff_segmentation algorithm will run with the defined parameters. 
 
-As a result of running the fff_segmentation algorithm with default parameters, two new files were generated in the ```Segmentation results``` folder, which was created in the Matlab's current path. The first, named [Points segmentation results Acoustic_signal](<Segmentation results/Points segmentation results Acoustic_signal.mat>), holds the results of the segmentation in the index format. The output data file name follows a definition that is based on the `Segmentation mode` choice, in this case `Points`, and the identification of the sensor signal in the dataset, in this case `Acoustic_signal`.   
[…]
