+++ readme.md
@@ -78,31 +78,30 @@ After running the fff_segmenter algorithm, a open a file explorer window. The us
 For the purpose of this example, the user can navigate to the `Data` folder and select the `Test1.mat` file.
 
 
-<img src="Example/Octave%file%navigation.png" alt="Figure 6 - Octave file selection" width="420">
+<img src="Example/Octave%20file%20navigation.png" alt="Figure 6 - Octave file selection" width="420">
 
 After selecting the data file, the user will be provided with the graphical interface displayed below. The filled values in the input fields are the default values for the `Segmentation mode` and `Unit` segmentation options, and with all of the available outputs toggled to `No`.
 
 <img src="Example/Octave%20Input%20GUI.png" alt="Figure 7 - Octave input GUI" width="420">
 
-
 The user will fill the fields with the values in regard to the `Test1.mat` dataset, with default values for the `Segmentation mode` and `Unit` segmentation options, and with all of the available outputs toggled to `Yes`.
 
-<img src="Example/Octave%20Input%20GUI%filled.png" alt="Figure 8 - Octave input GUI filled" width="420">
+<img src="Example/Octave%20Input%20GUI%20filled.png" alt="Figure 8 - Octave input GUI filled" width="420">
 
 When the the user press `Enter` on the keyboard, the graphical input interface will be closed and the fff_segmentation algorithm will run with the defined parameters. 
 
 As a result of running the fff_segmentation algorithm with default parameters, two new files were generated in the ```Segmentation results``` folder, which was created in the Octave's current path. The first, named `Points segmentation results Acoustic_signal`, holds the results of the segmentation in the `Points` segmentation choice, saved in struct format. The output data file name follows a definition that is based on the `Segmentation mode` choice, in this case `Points`, and the identification of the sensor signal in the dataset, in this case `Acoustic_signal`.   
 The figure below demonstrate the contents of the `Points segmentation results Acoustic_signal` in the Octave workspace. 
 
-<img src="Example/Octave%Segmentation%20index%20mode%20workspace.png" alt="Figure 9 - Points segmentation results in the workspace" width="420">
+<img src="Example/Octave%20Segmentation%20index%20mode%20workspace.png" alt="Figure 9 - Points segmentation results in the workspace" width="420">
 
 Opening each struct in Matlab, it is possible to observe that the point segmentation mode generates three columns for each geometric feature, and two columns for the pattern's separation. In regard to the geometric features tables, the first column is the ```Duration``` of the feature fabrication, the second column is the first instant index, identified as ```StartPoint```, of the feature fabrication, and the third column is the ```EndPoint``` of the feature fabrication in the default number of samples mode.
 
-![Figure 4 - Points segmentation results](Example/Octave%Segmentation%20index%20mode%20results.png)
+![Figure 10 - Points segmentation results](Example/Octave%20Segmentation%20index%20mode%20results.png)
 
 The second file, named `Segmentation results Acoustic_signal`, is the automatically saved figure with the predetermined resolution and formatting. The image obtained for this example is represented in the figure below.
 
-![Figure 5 - Saved figure](Example/Octave%Segmentation%20results%20'Test1'.png)
+![Figure 11 - Saved figure](Example/Octave%Segmentation%20results%20'Test1'.png)
 
 
 </details>
@@ -111,4 +110,4 @@ The second file, named `Segmentation results Acoustic_signal`, is the automatica
 # Application example
[…]
