### 2024-06-27 a84c823
Verify figure
