### 2024-06-27 b477fca
Verify figures with altered names
