### 2024-06-27 df155dc
Alter figures to insert percentage symbols
