+++ readme.md
@@ -40,11 +40,11 @@ fff_segmenter
 ```
 After running the fff_segmenter algorithm, the user will be provided with the graphical interface displayed below.
 
-<img src="Example/Matlab%20input%20GUI.png" alt="Figure 1 - Input interface" width="420">
+<img src="Example/Matlab%20Input%20GUI.png" alt="Figure 1 - Input interface" width="420">
 
 For the purpose of this example, the `Select data` button will be used to load the `Test1.mat` file from the `Data` folder. The `Select data` button will open a file explorer window, where the user can navigate to the `Data` folder and select the `Test1.mat` file. The rest of the input fields will be filled with the values in regard to the `Test1.mat` dataset, with default values for the `Segmentation mode` and `Unit` segmentation options, and with all of the available outputs toggled to `Yes`.
 
-<img src="Example/Matlab%20input%20GUI%20filled.png" alt="Figure 2 - Input interface filled with values" width="420">
+<img src="Example/Matlab%20Input%20GUI%20filled.png" alt="Figure 2 - Input interface filled with values" width="420">
 
 
 When the user toggle to `On` the `Run the segmentation` button, the graphical input interface will be closed and the fff_segmentation algorithm will run with the defined parameters. 
@@ -52,15 +52,15 @@ When the user toggle to `On` the `Run the segmentation` button, the graphical in
 As a result of running the fff_segmentation algorithm with default parameters, two new files were generated in the ```Segmentation results``` folder, which was created in the Matlab's current path. The first, named `Points segmentation results Acoustic_signal`, holds the results of the segmentation in the `Points` segmentation choice, saved in table format. The output data file name follows a definition that is based on the `Segmentation mode` choice, in this case `Points`, and the identification of the sensor signal in the dataset, in this case `Acoustic_signal`.   
 The figure below demonstrate the contents of the [Points segmentation results Acoustic_signal](<Segmentation results/Points segmentation results Acoustic_signal.mat>) in the Matlab workspace. 
 
-<img src="Example/Matlab%Segmentation%20index%20mode%20workspace.png" alt="Figure 3 - Points segmentation results in the workspace" width="420">
+<img src="Example/Matlab%20Segmentation%20index%20mode%20workspace.png" alt="Figure 3 - Points segmentation results in the workspace" width="420">
 
 Opening each table in Matlab, it is possible to observe that the point segmentation mode generates three columns for each geometric feature, and two columns for the pattern's separation. In regard to the geometric features tables, the first column is the ```Duration``` of the feature fabrication, the second column is the first instant index, identified as ```StartPoint```, of the feature fabrication, and the third column is the ```EndPoint``` of the feature fabrication in the default number of samples mode.
 
-![Figure 4 - Points segmentation results](Example/Matlab%Segmentation%20index%20mode%20results.png)
+![Figure 4 - Points segmentation results](Example/Matlab%20Segmentation%20index%20mode%20results.png)
 
 The second file, named `Segmentation results Acoustic_signal`, is the automatically saved figure with the predetermined resolution and formatting. The image obtained for this example is represented in the figure below.
 
-![Figure 5 - Saved figure](Example/Matlab%Segmentation%20results%20'Test1'.png)
+![Figure 5 - Saved figure](Example/Matlab%20Segmentation%20results%20'Test1'.png)
 
 </details>
 
@@ -78,18 +78,18 @@ After running the fff_segmenter algorithm, a open a file explorer window. The us
 For the purpose of this example, the user can navigate to the `Data` folder and select the `Test1.mat` file.
 
 
-<img src="example/Octave%20file%20navigation.png" alt="Figure 6 - Octave file selection" width="420">
[…]
