### 2024-06-27 217c485
Adjust figure identification
