+++ readme.md
@@ -83,13 +83,13 @@ For the purpose of this example, the user can navigate to the `Data` folder and 
 After selecting the data file, the user will be provided with the graphical interface displayed below. The filled values in the input fields are the default values for the `Segmentation mode` and `Unit` segmentation options, and with all of the available outputs toggled to `No`.
 
 
-<img src="example/Octave%20Input%20GUI.png" alt="Figure 7 - Octave input GUI" width="420">
+<img src="Example/Octave%20Input%20GUI.png" alt="Figure 7 - Octave input GUI" width="420">
 
 
 The user will fill the fields with the values in regard to the `Test1.mat` dataset, with default values for the `Segmentation mode` and `Unit` segmentation options, and with all of the available outputs toggled to `Yes`.
 
 
-<img src="example/Octave%20Input%20GUI%20filled.png" alt="Figure 8 - Octave input GUI filled" width="420">
+<img src="Example/Octave%20Input%20GUI%20filled.png" alt="Figure 8 - Octave input GUI filled" width="420">
 
 
 When the the user press `Enter` on the keyboard, the graphical input interface will be closed and the fff_segmentation algorithm will run with the defined parameters. 
