### 2024-06-27 dfb3d54
Alter figures width
