+++ .github/ISSUE_TEMPLATE/documentation_update.md
@@ -1,0 +1,17 @@
+---
+name: Documentation Update
+about: Suggest updates or improvements to the documentation
+title: "[DOCS] "
+labels: documentation
+assignees: ''
+
+---
+
+**Description of the update:**
+- Describe the part of the documentation that needs updating.
+
+**Reason:**
+- Explain why this update is necessary.
+
+**Additional context:**
+- Provide any additional context or details related to the documentation update.
