+++ .github/CONTRIBUTING.md
@@ -1,0 +1,64 @@
+# Contributing to FFF-Line-Segmentation
+
+We welcome contributions from everyone. By participating in this project, you agree to abide by our Code of Conduct. Here are some guidelines to help you get started.
+
+## How to Contribute
+
+1. Fork the Repository
+- Navigate to the `FFF-Line-Segmentation` repository on GitHub.
+- Click the "Fork" button in the upper right corner of the page.
+
+
+2. Create a Branch
+- Clone your fork to your local machine:
+```sh
+git clone https://github.com/your-username/FFF-Line-Segmentation.git
+```
+
+- Create a new branch for your changes:
+```sh
+git checkout -b your-branch-name
+```
+
+3. Make Your Changes
+- Make the necessary changes in your branch.
+- Ensure your code follows the project's coding style and conventions.
+- Test your changes to make sure they work as expected with all of the datasets available in [Data](/data/).
+
+4. Commit Your Changes
+- Commit your changes with a descriptive commit message:
+
+```sh
+git commit -m "Description of the changes made"
+```
+
+5. Push to Your Fork
+- Push your changes to your forked repository:
+```sh
+git push origin your-branch-name
+```
+
+6. Create a Pull Request
+- Navigate to the original repository on GitHub.
+- Click the "New pull request" button.
+- Select the branch you created from your fork and create the pull request.
+- Provide a clear and descriptive title and description for your pull request.
+
+## Reporting Issues
+
+If you encounter a bug or have a feature request, please create an issue. Before submitting a new issue, please check the list of open issues to see if the issue has already been reported.
+
+### How to Report an Issue
+
+- Check if an issue has already been reported. Browse the list of open issues to see if someone has already reported the problem.
+- Create a new issue if the problem is not listed. Follow the [issue template](.github/ISSUE_TEMPLATE) and provide as much detail as possible.
+- Be patient. The maintainers will review the issue and respond as soon as possible.
+
+## For questions or support:
+
+- Check the documentation. Many common issues and questions are addressed in the project's [documentation](/Documentation.pdf).
+- Search closed issues. Someone else might have had the same problem or question.
+- Open a new issue if you cannot find an answer. Be clear and specific in your question or problem description.
+- Participate in discussions. Others in the community may have helpful solutions or suggestions.
+
+By contributing to this project, you agree to follow our [Code of Conduct](.github/CODE_OF_CONDUCT.md).
