+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,32 @@
+---
+name: Bug Report
+about: Create a report to help us improve
+title: "[BUG] "
+labels: bug
+assignees: ''
+
+---
+
+**Description of the bug:**
+- Describe the bug in a concise and clear manner.
+
+**Steps to Reproduce:**
+1. Step 1
+2. Step 2
+3. Step 3
+
+**Expected Behavior:**
+- Describe what you expected to happen.
+
+**Actual Behavior:**
+- Describe what actually happened.
+
+**Screenshots:**
+- If applicable, add screenshots to help explain your problem.
+
+**Environment:**
+- OS: [e.g., Windows 10, Ubuntu 20.04]
+- Software version: [e.g., v1.0.0]
+
+**Additional context:**
+- Add any other context about the problem here.
