+++ .github/ISSUE_TEMPLATE/question.md
@@ -1,0 +1,14 @@
+---
+name: Question
+about: Ask a question about this project
+title: "[QUESTION] "
+labels: question
+assignees: ''
+
+---
+
+**Question:**
+- Clearly state your question.
+
+**Additional context:**
+- Provide any additional context or details related to your question.

[…]
