### 2024-06-28 e4becbd
Alter the documentation pdf file and the main paper
