+++ fff_segmenter.m
@@ -25,7 +25,7 @@ function fff_segmenter
 %
 % Outputs:
 % - The main outputs of the function are the segmentation results, which can be
-%   the segmentation index (points), in seconds or in or the segments of signal (segments), based
+%   the segmentation index (points), in seconds or in or the segments of signal (Number of samples), based
 %   on the user's choice. These segmentation results are stored in the user current
 %   matlab/octave workspace.
 % - Additionally, if the user opted for it, the segmentation results can be automatically
@@ -39,7 +39,7 @@ function fff_segmenter
 % <a href="matlab: disp('adjust_internal: Adjust the internal pattern segmentation results for a specific scenario. Please read the comments under the subfunction for further details.') ">adjust_internal</a>
 % <a href="matlab: disp('cleanUpWS: Clean up the base workspace from the fff_segmenter variables.') ">cleanUpWS</a>
 % <a href="matlab: disp('Compos3r: Compose a signal based on the original acoustic signal and on specific segmentation indexes. Please read the comments under the subfunction for further details.') ">Compos3r</a>
-% <a href="matlab: disp('convUni: convUni Obtain a time (s) value based on a number of samples value, a reference signal, and on the sampling frequency. Please read the comments under the subfunction for further details.') ">convUni</a>
+% <a href="matlab: disp('convUni: convUni Obtain a time (s) value based on a Number of samples value, a reference signal, and on the sampling frequency. Please read the comments under the subfunction for further details.') ">convUni</a>
 % <a href="matlab: disp('detectAnom: Detect the anomaly verified in the Test1.mat dataset. Please read the comments under the subfunction for further details.') ">detectAnom</a>
 % <a href="matlab: disp('determin_separation_point: Determine the separation point based on the duration matrix and on a reference duration value. Please read the comments under the subfunction for further details.') ">determin_separation_point</a>
 % <a href="matlab: disp('gen_graph: Generate the graphical visualization of the segmentation results. Please read the comments under the subfunction for further details.') ">gen_graph</a>
@@ -50,7 +50,7 @@ function fff_segmenter
 % <a href="matlab: disp('Normaliz3r: Normalize the amplitude of a signal between -1 and 1. Please read the comments under the subfunction for further details.') ">Normaliz3r</a>
 % <a href="matlab: disp('obtainDuration: Obtain the duration matrix based on the acoustic signal and on the step motor control signals. Please read the comments under the subfunction for further details.') ">obtainDuration</a>
 % <a href="matlab: disp('obtainOutput: Obtain a formated output with the segmentation results. Please read the comments under the subfunction for further details.') ">obtainOutput</a>
-% <a href="matlab: disp('obtain_time_vec: Obtain a time (s) vector based on the signal's number of samples length and on the sampling frequency. Please read the comments under the subfunction for further details.') ">obtain_time_vec</a>
+% <a href="matlab: disp('obtain_time_vec: Obtain a time (s) vector based on the signal's Number of samples length and on the sampling frequency. Please read the comments under the subfunction for further details.') ">obtain_time_vec</a>
 % <a href="matlab: disp('saveFig: Save the graphical visualization of the segmentation results in a predefined format and resolution. Please read the comments under the subfunction for further details.') ">saveFig</a>
[…]
