### 2024-06-28 thiago-glissoi/FFF-Line-Segmentation#4 @thiago-glissoi
Hey @olivecha,

Thanks for contributing with this issue!

We've updated the [Documentation](https://github.com/thiago-glissoi/FFF-Line-Segmentation/blob/main/Documentation.pdf) to address your feedback.

Specifically, we created a section called **Acquirement of the Necessary Signals** where the required information is now available.

If you need more details or have any other suggestions, feel free to let us know.

Best,

Thiago
