### 2024-06-28 thiago-glissoi/FFF-Line-Segmentation#5 @thiago-glissoi
@olivecha 👋

Thanks for the issue.

We've made significant changes to how users input the necessary data, signal names, and segmentation options. Upon execution, the `fff_segmenter` script now displays graphical interfaces for user input, available for both MATLAB and Octave.

Please check if these changes meet the requirements.

Best,

Thiago
