### 2024-06-28 thiago-glissoi/FFF-Line-Segmentation#6 @thiago-glissoi
Hey @olivecha  👋

Thanks for bringing this up!

We've created a [validation function](https://github.com/thiago-glissoi/FFF-Line-Segmentation/blob/main/Validate_segmentation.m) to help users verify unexpected changes in behavior due to different MATLAB/Octave versions or operating systems.

This [validation function](https://github.com/thiago-glissoi/FFF-Line-Segmentation/blob/main/Validate_segmentation.m) compares results with the [Test1.mat data results](https://github.com/thiago-glissoi/FFF-Line-Segmentation/blob/main/Segmentation%20results/points%20segmentation%20results%20Test1.mat) and should be run with the user's data of interest.

Best,

Thiago
