### 2024-06-28 thiago-glissoi/FFF-Line-Segmentation#2 @thiago-glissoi
Hey @olivecha

Thanks for the issue!

To address your request, we've added the required function documentation strings and detailed explanations for every argument, return values, and the function's purpose. We've also updated the [Documentation](https://github.com/thiago-glissoi/FFF-Line-Segmentation/blob/main/Documentation.pdf) to increase content, and also improve understanding for the community.

If any further changes are needed, feel free to let us know.

Best,
Thiago
