### 2024-06-28 #6620 @thiago-glissoi
Hey @bmcfee  and @olivecha 👋

The authors have completed the work with the main and secondary issues brought by @olivecha

> As can be seen from the checklist, some work on the documentation should be done before publication, but the rest is in order. I don't think elaborate packaging should be expected from MATLAB projects, especially as dependencies are a non-issue. However, I don't think the project makes use of any MATLAB specific functions or toolboxes and could be rewritten with `scipy` and `numpy`, something the author should consider.
>
> The main issues requiring a certain amount of work that must be resolved before publication are:
>
> * [Code comments and documentation  thiago-glissoi/FFF-Line-Segmentation#2](https://github.com/thiago-glissoi/FFF-Line-Segmentation/issues/2)
> * [Data structure for example usage thiago-glissoi/FFF-Line-Segmentation#4](https://github.com/thiago-glissoi/FFF-Line-Segmentation/issues/4)
> * [User and `fff_segmenter` inputs thiago-glissoi/FFF-Line-Segmentation#5](https://github.com/thiago-glissoi/FFF-Line-Segmentation/issues/5)
> * [Add at least one test thiago-glissoi/FFF-Line-Segmentation#6](https://github.com/thiago-glissoi/FFF-Line-Segmentation/issues/6)

The authors greatly appreciate the recommendations made by the reviewer.
