### 2024-07-01 thiago-glissoi/FFF-Line-Segmentation#2 @olivecha
Could you mention the issue number (#2) in the commit messages of the changes addressing the issue ?
(Same for all issues, so we can track what was done to remediate the issues)
