### 2024-07-01 thiago-glissoi/FFF-Line-Segmentation#2 @thiago-glissoi
@olivecha Sure!

There were some alterations to address the issues that were made in multiple commits. I will insert the comment in each of them.
