### 2024-07-01 thiago-glissoi/FFF-Line-Segmentation#4 @olivecha
Thanks for the clarification. Maybe there is still a little gap on how the raw data was processed into Matlab files or perhaps this is handled by your signal acquisition tools? Lets say I have a CSV with columns "time", "DirX on/off", "DirY on/off" and a WAV file with a sample rate of 32 kHz how could I input that into your tool ?
Also mention that the DirX/Y data is in every Test.mat file that was not clear in the documentation.
