### 2024-07-01 thiago-glissoi/FFF-Line-Segmentation#4 @thiago-glissoi
> Thanks for the clarification. Maybe there is still a little gap on how the raw data was processed into Matlab files or perhaps this is handled by your signal acquisition tools? Lets say I have a CSV with columns "time", "DirX on/off", "DirY on/off" and a WAV file with a sample rate of 32 kHz how could I input that into your tool ? Also mention that the DirX/Y data is in every Test.mat file that was not clear in the documentation.

To directly answer your question, the raw data 'processing' into Matlab files was conducted in the example datasets by the Data Acquisition Device (DAQ).
As mentioned in the updated documentation, for all three of the example datasets, the DAQ used was the DL 850 oscilloscope from Yokogawa®, Tokyo, Japan. This DAQ present the option to save the waveform data directly in MATLAB format ( for reference, view page 16-10 of the [DL850 user manual](https://www.yokogawa.com/pdf/provide/E/GW/IM/0000028031/0/IMDL850E-02EN.pdf) ).

The `fff_segmenter` script requires that you possess a single .mat file that holds:

1. The sensor signal that will be segmented;
2. The X axis direction signal;
3. The Y-axis direction signal.

These signals, when the .mat is loaded to the MATLAB/Octave workspace, needs to be in `double` type one line multiple columns values. Also, the three signals need to be obtained synchronously (that is, at the start and end of the printing process), and at the same Sampling Frequency (Fs). This will guarantee that the three signals possess the same column numbers.

In the hypothetical situation you mentioned, an approach you could try is to use MATLAB/Octave input options to convert upon insertion to the workspace the original data structure in order to obtain three separate `double` type one line multiple columns variables.

I have updated the documentation to insert the information about the example dataset DAQ output, alongside the information that the DIRX/Y data is in every Test.mat file.
