### 2024-07-01 71fdfe4
Correct minor typos in the readme
