### 2024-07-01 bf9fc1d
Correct minor typo in the Contributing.md
