### 2024-07-02 thiago-glissoi/FFF-Line-Segmentation#2 @olivecha
This is much better, the documentation of the sub-functions is a bit hard to get to, but I guess this is a Matlab thing.

You could mention when (important) sub-functions are called in the three steps of the main function instead of naming then alphabetically, as this would make it easier to understand their utility, but this would not be necessary for publication.
