### 2024-07-02 thiago-glissoi/FFF-Line-Segmentation#6 @olivecha
I looked at and ran the validation function, and it does not behave like a software test...
The tests should not be run against arbitrary data (well they could to be more robust, but simpler tests would need to be run first).
For your project, a simple test would be to run the main function and compare with known output:
```matlab
load("Data/Test1.mat");
fff_segmenter(Acoustic_signal,Dir_X,Dir_Y,200e3);
% Load the result
load("Segmentation results/points segmentation results Test1.mat")
% Create a copy of the Segmentation results in another folder
% like "test_data/"
load("test_data/Points segmentation results test.mat")
% Do this for each result array
test1 = all(abs(result_contour(:) - test_contour(:)) < 1e-10);
if test1
    disp('Contour match test reference')
else
    disp('Test failed when comparing contour')
end
```

This would allow you to validate that changes in the code don't break the existing behaviour (like when accepting pull requests) and that the environments of other users provides the expected functions (various Matlab/Octave versions).
