### 2024-07-02 thiago-glissoi/FFF-Line-Segmentation#5 @olivecha
This is nice, but would require a little more documentation, as it was really not clear what the fields are for a new user.
1. What format is expected in the "Select Data" file loader
2. That the sampling frequency is required. This is a bit uncommon, as most (if not all) audio formats store the sampling frequency in the metadata.
3. That the "Identification" fields are names of the associated variables in the .mat file. It would really help to have the variables names in the file in a drop-down menu.
4. Add a more detailed description of what the segmentation mode and unit choice do in the "How to use" section
5. If you run the App with the default "No" values, there is no output?

Adding annotations with arrows in the graphical interface figure could do the trick
