### 2024-07-02 thiago-glissoi/FFF-Line-Segmentation#4 @olivecha
That is more clear. It does make sense that a Matlab script takes Matlab files as an input.
