### 2024-07-02 #6620 @olivecha
Only two issues remain:
https://github.com/thiago-glissoi/FFF-Line-Segmentation/issues/5
https://github.com/thiago-glissoi/FFF-Line-Segmentation/issues/6
Which will meet the "Example usage" and "Automated tests" checks respectively
