### 2024-07-02 e1977b9
Alter the validation function
