+++ Validate_segmentation.m
@@ -1,1724 +1,80 @@
-function Validate_segmentation
+function Validate_segmentation(sensorSignal, dirX, dirY, Fs)
 %   ATTENTION!
 %
-%   THIS IS NOT the main segmentation function. This is a validation script that should be run when
+%   THIS IS NOT the main segmentation function. This is a validation function that should be run when
 %   a verification is required to confirm that the segmentation is being performed correctly and there is
 %   no different behaviours due to different Matlab/Octave versions or operating systems.
 %   
-%   The main segmentation function is FFF_Line_Segmentation.m
+%   The main segmentation function is fff_segmenter.m
 %
+% Validate_segmentation  This validation function run the main function and compare with known output.
 %
-%   Run this function by calling Validate_segmentation in the command window.
-%   the validation function will run in a similar manner to that of the main function, but without any outputs.
+%   Validate_segmentation(sensorSignal, dirX, dirY, Fs)
+% 
+% Inputs:
+% - sensorSignal (array): The acoustic signal obtained from the FFF process.
+% - dirX (array): The signal obtained from the X-axis step motor control.
+% - dirY (array): The signal obtained from the Y-axis step motor control.
+% - Fs (double): The sampling frequency of the signals.
+%
+%#ok<*LOAD>
+%#ok<*USENS>
+
 disp (' ');
 disp ('ATTENTION!');
 disp ('This is a validation script that should be run when a verification is required to confirm ');
 disp ('that the segmentation is being performed correctly and there is no different behaviours due ');
 disp ('to different Matlab/Octave versions or operating systems.');
-disp (' ');
-disp ('Run it in the same way as the main function, but without any outputs.');
 disp ('/ATTENTION!');
 disp (' ');
 
-if exist('OCTAVE_VERSION', 'builtin') == 0
-    % Is not running in octave
-    % APP User Inputs
-    app = APP_User_Inputs;
-    while strcmp(app.RunthesegmentationSwitch.Value,'Off') == 1
-        pause(0.5); % In order to avoid missuse of the CPU
-    end
-
-    % Verify presence of identifications
-    signalIdentifier_exists = evalin('base', ['exist(''', 'signalIdentifier', ''', ''var'')']);
-    DirXidentification_exists = evalin('base', ['exist(''', 'DirXidentification', ''', ''var'')']);
-    DirYidentification_exists = evalin('base', ['exist(''', 'DirYidentification', ''', ''var'')']);
-
-    if signalIdentifier_exists == 1 && DirXidentification_exists == 1 &&...
-            DirYidentification_exists == 1
-    else
-        disp ('ATTENTION!')
-        disp ('A necessary signal identification was not provided.');
-        disp ('Please, run the function again and provide the proper signal identifications.');
-        delete(app.UIFigure);
-        cleanUpWS;
-        return;
-    end    
-
-    % Verify Sampling frequency
-    Fs_exists = evalin('base', ['exist(''', 'Fs', ''', ''var'')']);
-    if Fs_exists == 1
-        Fs = evalin('base', 'Fs');
-    else
-        disp ('ATTENTION!')
-        disp ('No numerical value attributed to Sampling frequency (Fs) field.');
-        disp ('Please, run the function again and provide the proper Fs value.');
-        delete(app.UIFigure);
-        cleanUpWS;
-        return;
-    end
-
[…]
