### 2024-07-02 thiago-glissoi/FFF-Line-Segmentation#6 @thiago-glissoi
@olivecha

Understood.
I have updated the [validation function](https://github.com/thiago-glissoi/FFF-Line-Segmentation/blob/main/Validate_segmentation.m) to follow the type of test that you proposed.

Cheers,
Thiago
