+++ readme.md
@@ -40,6 +40,21 @@ fff_segmenter
 ```
 After running the fff_segmenter algorithm, the user will be presented with the graphical interface displayed below.
 
+In the graphical interface:
+- The user must use the `Select data` button to select the .mat file which contains the acoustic signal and the X and Y direction control signals.
+- The user must insert the `Sampling frequency` of the signals in Hertz (Hz).
+- The user must insert the name of the sensor signal that exists in the selected .mat file on the `Raw signal identification` field.
+- The user must insert the name of the direction X axis signal that exists in the selected .mat file on the `DirX identification` field.
+- The user must insert the name of the direction Y axis signal that exists in the selected .mat file on the `DirY identification` field.
+- The user must select the `Segmentation mode` between `Points` and `Segments`. If the choice is not made, the default `Points` mode will be utilized.
+- The user must select the `Unit` segmentation between `Number of samples` and `Seconds`. If the choice is not made, the default `Number of samples` mode will be utilized.
+- The user may select additional `Outputs` that can be generated by the algorithm. The available outputs are `Obtain graphical visualization`, `Save graphical visualization if obtained`, and `Obtain automatic .mat files`. If none additional `Outputs` are selected, the `fff_segmenter` function output will only be the segmentation results to the current Matlab workspace.
+
+About the additional outputs:
+- The `Obtain graphical visualization` option will generate a figure with the segmentation results displayed on top of the `Sensor signal`.
+- The `Save graphical visualization if obtained` option will save the generated figure in the `Segmentation results` folder.
+- The `Obtain automatic .mat files` option will save the segmentation results in the `Segmentation results` folder in .mat format. The output data file name follows a definition that is based on the `Segmentation mode` choice, and the identification of the sensor signal in the dataset.
+
 <img src="Example/Matlab%20Input%20GUI.png" alt="Figure 1 - Input interface" width="420">
 
 For the purpose of this example, the `Select data` button will be used to load the `Test1.mat` file from the `Data` folder. The `Select data` button will open a file explorer window, where the user can navigate to the `Data` folder and select the `Test1.mat` file. The rest of the input fields will be filled with the values in regard to the `Test1.mat` dataset, with default values for the `Segmentation mode` and `Unit` segmentation options, and with all of the available outputs toggled to `Yes`.
@@ -74,13 +89,31 @@ Run the `fff_segmenter` script by typing directly into the command window follow
 ```Octave
 fff_segmenter
 ```
-After running the fff_segmenter algorithm, a file explorer window will appear. For the purpose of this example, the user can navigate to the `Data` folder and select the `Test1.mat` file.
+After running the fff_segmenter algorithm, a file explorer window will appear. 
+
+- The user must select the .mat file which contains the acoustic signal and the X and Y direction control signals.
+
+For the purpose of this example, the user can navigate to the `Data` folder and select the `Test1.mat` file.
 
 
 <img src="Example/Octave%20file%20navigation.png" alt="Figure 6 - Octave file selection" width="420">
 
 After selecting the data file, the user will be presented with the graphical interface displayed below. The filled values in the input fields are the default values for the `Segmentation mode` and `Unit` segmentation options, and with all of the available fields set to `N`.
 
+In the graphical interface:
+
+- The user must insert the name of the sensor signal that exists in the selected .mat file on the `Sensor signal identification` field.
+- The user must insert the name of the direction X axis signal that exists in the selected .mat file on the `X-axis signal identification` field.
+- The user must insert the name of the direction Y axis signal that exists in the selected .mat file on the `Y-axis signal identification` field.
+- The user must insert the `Sampling frequency` of the signals in Hertz (Hz).
[…]
