### 2024-07-02 thiago-glissoi/FFF-Line-Segmentation#5 @thiago-glissoi
@olivecha  👋

> 1. What format is expected in the "Select Data" file loader

The information was inserted through an annotation on the graphical interface figure and to the `README.md` document.

> 2. That the sampling frequency is required. This is a bit uncommon, as most (if not all) audio formats store the sampling frequency in the metadata.

The information was inserted in the `README.md` document.

> 3. That the "Identification" fields are names of the associated variables in the .mat file. It would really help to have the variables names in the file in a drop-down menu.

The information was inserted through an annotation on the graphical interface figure and to the `README.md` document.

> 4. Add a more detailed description of what the segmentation mode and unit choice do in the "How to use" section

The information was inserted in the `README.md` document.

> 5. If you run the App with the default "No" values, there is no output?

If you run the `fff_segmenter` script with default values, that is with "No" values, there still will be the segmentation results output to the current MATLAB/Octave workspace.  The information was inserted in the `README.md` document.

> Adding annotations with arrows in the graphical interface figure could do the trick

Annotations had been inserted in the graphical interface figure.

If additional informations are required, please let me know.

Best,
Thiago
