### 2024-07-02 #6620 @thiago-glissoi
@olivecha

> Only two issues remain:
>
> * [User and `fff_segmenter` inputs thiago-glissoi/FFF-Line-Segmentation#5](https://github.com/thiago-glissoi/FFF-Line-Segmentation/issues/5)
> * [Add at least one test thiago-glissoi/FFF-Line-Segmentation#6](https://github.com/thiago-glissoi/FFF-Line-Segmentation/issues/6)
>   Which will meet the "Example usage" and "Automated tests" checks respectively

I have completed the work in this issues.

The authors greatly appreciate the recommendations made by the reviewer.
