### 2024-07-02 5c55827
Correct minor typos in the readme.md about the user input on octave
