+++ readme.md
@@ -108,12 +108,12 @@ In the graphical interface:
 - The user must insert the `Sampling frequency` of the signals in Hertz (Hz).
 - The user must select the `Segmentation mode` between `Points` and `Segments`. If the choice is not made, the default `Points` mode will be utilized.
 - The user must select the `Unit` segmentation between `Number of samples` and `Seconds`. If the choice is not made, the default `Number of samples` mode will be utilized.
-- The user may select additional `Outputs` that can be generated by the algorithm. The available outputs are `Obtain graphical visualization`, `Save graphical visualization if obtained`, and `Obtain automatic .mat files`. If none additional `Outputs` are selected, the `fff_segmenter` function output will only be the segmentation results to the current Octave workspace. 
+- The user may select additional `Outputs` that can be generated by the algorithm. The available outputs are `Generate figure`, `Save figure`, and `Auto save files`. If none additional `Outputs` are selected, the `fff_segmenter` function output will only be the segmentation results to the current Octave workspace. 
 
 About the additional outputs:
-- The `Obtain graphical visualization` option will generate a figure with the segmentation results displayed on top of the `Sensor signal`.
-- The `Save graphical visualization if obtained` option will save the generated figure in the `Segmentation results` folder.
-- The `Obtain automatic .mat files` option will save the segmentation results in the `Segmentation results` folder in .mat format. The output data file name follows a definition that is based on the `Segmentation mode` choice, and the identification of the sensor signal in the dataset.
+- The `Generate figure` option will generate a figure with the segmentation results displayed on top of the `Sensor signal`.
+- The `Save figure` option will save the generated figure in the `Segmentation results` folder.
+- The `Auto save files` option will save the segmentation results in the `Segmentation results` folder in .mat format. The output data file name follows a definition that is based on the `Segmentation mode` choice, and the identification of the sensor signal in the dataset.
 
 <img src="Example/Octave%20input%20GUI.png" alt="Figure 7 - Octave Input GUI" width="420">
 
