### 2024-07-02 9287bec
Alter slightly the Validate_segmentation functionality
