### 2024-07-02 thiago-glissoi/FFF-Line-Segmentation#6 @olivecha
Closing as this is solved
