### 2024-07-02 thiago-glissoi/FFF-Line-Segmentation#5 @olivecha
Closing as this is acceptable
