### 2024-07-04 #6620 @DARSakthi
Hello

I apologise for the delays in my response but I have been travelling most of the last two months

I concur with @olivecha's comments and since they have been addressed I do not have much to add.  I have all but completed the checklist.  I have the following two requests:

The first citation looks incomplete --- the paper belongs to the proceedings of some particular conference no?

For the reader's context can you comment more on how the data in https://github.com/thiago-glissoi/FFF-Line-Segmentation/tree/main/Data was obtained?  Some machine settings or collection parameters or something of that nature?  A readme in that folder ought to be fine I think (or perhaps this information is already available somewhere else and I have simply not seen it?)

Cheers

DS
