### 2024-07-04 #6620 @thiago-glissoi
Hello @DARSakthi

Thank you for your help.

In regard to your requests.

> The first citation looks incomplete --- the paper belongs to the proceedings of some particular conference no?

Yes, the first citation is from a conference. Initially, the article is published in the conference proceedings, and later, it is republished in a journal using the same DOI, with updated metadata. You can find the full article and its details at: https://www.mdpi.com/2504-3900/69/1/40

> For the reader's context can you comment more on how the data in https://github.com/thiago-glissoi/FFF-Line-Segmentation/tree/main/Data was obtained? Some machine settings or collection parameters or something of that nature? A readme in that folder ought to be fine I think (or perhaps this information is already available somewhere else and I have simply not seen it?)

I believe that this was already addressed in a previous [issue](https://github.com/thiago-glissoi/FFF-Line-Segmentation/issues/4). You can find details about how the data was obtained in the [Documentation](https://github.com/thiago-glissoi/FFF-Line-Segmentation/blob/main/Documentation.pdf). More specifically, in the section called **Acquirement of the Necessary Signals**. If more detailing is required, please, let me know.

Cheers,

Thiago
