### 2024-07-04 c99cac2
Update some articles information based on their doi metadata
