+++ paper.bib
@@ -1,18 +1,16 @@
-@article{Lopes2022,
-   abstract = {Additive manufacturing (AM) has been played a crucial role in the fourth industrial revolution. Sensor-based monitoring technologies are essential in detecting defects and providing feedback for process control. Acoustic emission (AE) sensors have been used for long time in a wide range of processes and fields, but they are still a challenge in AM processes. This work presents a study on the AE signals in the time-domain - raw and root mean square (RMS) values - regarding their behavior during the manufacturing of a single-layer part in the Fused Filament Fabrication process for two infill patterns. Tests were conducted in a Cartesian 3D printer using Polylactic Acid material. The AE sensor was attached to the printer table through a magnetic coupling, and the signal was collected by an oscilloscope at 1 MHz sampling frequency. It was found that the raw AE signals behaved quite differently not just for the two infill patterns, but within the same pattern. The raw and RMS AE signals contain many spikes along the whole process, but the higher ones were those generally occurring at the end and/or start of a fabrication line. The RMS values, however, was useful for finding the start and end times of each fabricated line for both patterns. The mean RMS values have shown a nearly constant but distinct averages for the only-extruder, only-table and extruder-table movements.},
+@inproceedings{Lopes2022,
    author = {Thiago Glissoi Lopes and Paulo Roberto Aguiar and Thiago Valle França and Pedro de Oliveira Conceição Júnior and Cristiano Soares Junior and Zaqueu Ricardo Fernando Antonio},
+   city = {Basel Switzerland},
    doi = {10.3390/ecsa-9-13285},
-   editor = {Stefano Mariani},
-   issue = {1},
-   journal = {Engineering Proceedings},
-   month = {1},
+   journal = {The 9th International Electronic Conference on Sensors and Applications},
+   month = {11},
    pages = {83},
    publisher = {MDPI},
    title = {Time-Domain Analysis of Acoustic Emission Signals during the First Layer Manufacturing in FFF Process},
-   volume = {27},
    url = {https://www.mdpi.com/2673-4591/27/1/83},
    year = {2022},
 }
+
 
 @article{Wu2015,
    author = {Haixi Wu and Yan Wang and Zhonghua Yu},
@@ -60,23 +58,19 @@
    year = {2017},
 }
 
-@article{Carmo2020,
[…]
