### 2024-07-12 #6620 @bmcfee
@thiago-glissoi I'm now working down the editorial tasks for the post-review phase; please see https://github.com/openjournals/joss-reviews/issues/6620#issuecomment-2225563820 for a list of author tasks and let me know as you complete them.

Meanwhile, I'm reading the text of the paper and have a couple of suggestions / corrections:

- Line 41: "orientated" → "oriented"
- Line 47: refers to Figure 2 which does not exist
