### 2024-07-12 b29497b
Correct typo and reference to a previously existent figure in the article
