+++ paper.md
@@ -55,9 +55,9 @@ A consequence of this programming approach is that the FFF signal segmentation i
 
 ![Segmentation example with `fff_segmenter`. (a) signal with raster lines; (b) signal with transition between raster lines](images/segmentationImage.png){width=100%}
 
-### Result orientated
+### Result oriented
 
-In addition to its automatic feature, `fff_segmenter` is result-oriented. Firstly, the segmentation results can be presented in graphical windows, such as the one presented in Figure 1, that allow for a quick visual inspection of the segmentation results. This feature allows for a quick and accurate evaluation of the script's performance. Secondly, the segmentation results can be autosaved in predefined workspace data formats. These formats allow for the easy import of segmentation results into other MATLAB analyses. Figure 2 reveals the input options for the process operator, and the obtained output information. 
+In addition to its automatic feature, `fff_segmenter` is result-oriented. Firstly, the segmentation results can be presented in graphical windows, such as the one presented in Figure 1, that allow for a quick visual inspection of the segmentation results. This feature allows for a quick and accurate evaluation of the script's performance. Secondly, the segmentation results can be autosaved in predefined workspace data formats. These formats allow for the easy import of segmentation results into other MATLAB analyses.
 Figure 1 shows the segmentation results obtained by applying the `fff_segmenter` script to an acoustic signal. It is possible to observe that the acoustic signal presents many periods that are not of interest for monitoring purposes and, therefore, are not segmented. Further analysis of Figure 1 allows for the identification of contour lines, raster lines, and the transition period between printing patterns. During the transition period between printing patterns, the printer extruder repositions itself without depositing filament. The observation of the transition period analysis window (a.1) in Figure 1(a) confirms an important property of the raster lines printing pattern, which is the fact that the duration of the lines increases, and the raster lines are always separated by a transition period that presents the same duration. These transitions between raster lines, presented in Figure 1(b) analysis window (b.1), are periods of the signal during which there is deposition of filament.
 Due to being developed with the part's feature in focus, the algorithms in fff_segmenter are more direct, maintainable, and, above all, easier to understand with respect to the printer deposition movements.
 
