### 2024-07-12 #6620 @thiago-glissoi
Hey @bmcfee 👋

I have correct these problems in [commit b29497b](https://github.com/thiago-glissoi/FFF-Line-Segmentation/commit/b29497b2300a550c2be4669aebc0f6c1e9974612).
> * Line 41: "orientated" → "oriented"
> * Line 47: refers to Figure 2 which does not exist

I have created a current version of the repository [v3.0.1](https://github.com/thiago-glissoi/FFF-Line-Segmentation/releases/tag/v3.0.1).

I also have uploaded this version to Zenodo on [10.5281/zenodo.12733091 ](https://doi.org/10.5281/zenodo.12733091).

I have verified both the article and references by using the editorialbot, and seems ok. If anything else is necessary, let me know.

Cheers
