### 2024-07-12 0bf558a
Alter the URL for the thesis citation to match the university repository page, not the thesis document itself
