+++ paper.bib
@@ -77,7 +77,7 @@
    institution = {Faculdade de Engenharia de Bauru (FEB) - Universidade Estadual Paulista (UNESP)},
    pages = {219},
    title = {Detecção e classificação de anomalias durante o processo de manufatura aditiva por fusão e deposição por meio de transdutores piezelétricos e processamento de sinais},
-   url = {https://repositorio.unesp.br/bitstream/handle/11449/204399/lopes_tg_me_bauru_par.pdf?sequence=5},
+   url = {http://hdl.handle.net/11449/204399},
    year = {2021},
 }
 
