### 2024-07-12 #6620 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
