### 2024-07-12 #6620 @danielskatz
👋 @thiago-glissoi - I'm the track editor for this submission, and I've just proofread it. I just have two requests:
1. Please change "cartesian" to "Cartesian".
2. Please add a period at the end of the Acknowledgements sentence.

After that, we can proceed to acceptance and publication.
