### 2024-07-12 #6620 @danielskatz
@thiago-glissoi - sorry, I also have one request that's not for the paper: can you update the metadata for the zenodo deposit to the change the title to match the title of the paper? This should not require a new version/DOI; it's just a metadata change.
