### 2024-07-12 edbde40
Correct some wording and pontuation
