+++ paper.md
@@ -63,11 +63,11 @@ Due to being developed with the part's feature in focus, the algorithms in fff_s
 
 ### Scalable
 
-The algorithms in `fff_segmenter` were developed with a focus on cartesian [RepRap](https://reprap.org/wiki/RepRap)-based FFF printers. This is due to the fact that cartesian [RepRap](https://reprap.org/wiki/RepRap)-based FFF printers are commonly used for research purposes [@Lopes2022; @Carmo2020; @Wu2015; @Liu2018b]. Additionally, the segmentation algorithms were developed to segment the contour and raster printing lines of rectangular-shaped monolayer parts. The STL file of the part was sliced to G-code following standard printing parameters provided by the printer manufacturer. The documentation of `fff_segmenter` provides a detailed explanation of the segmentation algorithms and how they were developed to address the particularities of the cartesian [RepRap](https://reprap.org/wiki/RepRap)-based FFF printer and the rectangular-shaped monolayer part, providing a scalable framework for the segmentation of parts with different shapes and printing parameters.
+The algorithms in `fff_segmenter` were developed with a focus on Cartesian [RepRap](https://reprap.org/wiki/RepRap)-based FFF printers. This is due to the fact that Cartesian [RepRap](https://reprap.org/wiki/RepRap)-based FFF printers are commonly used for research purposes [@Lopes2022; @Carmo2020; @Wu2015; @Liu2018b]. Additionally, the segmentation algorithms were developed to segment the contour and raster printing lines of rectangular-shaped monolayer parts. The STL file of the part was sliced to G-code following standard printing parameters provided by the printer manufacturer. The documentation of `fff_segmenter` provides a detailed explanation of the segmentation algorithms and how they were developed to address the particularities of the Cartesian [RepRap](https://reprap.org/wiki/RepRap)-based FFF printer and the rectangular-shaped monolayer part, providing a scalable framework for the segmentation of parts with different shapes and printing parameters.
 
 In summary, `fff_segmenter` fulfills the need in FFF process monitoring research for a signal segmentation script that is automatic, result-oriented, and scalable, as well as providing a framework for the community to scale the script to cater to other FFF printers and parts.
 
 # Acknowledgements
-The authors are thankful to the National Council for Scientific and Technological Development (CNPq) for supporting this research under the Grant # 306774/2021-6, and to the Coordenação de Aperfeiçoamento de Pessoal de Nível Superior – Brasil (CAPES) – Finance Code 001
+The authors are thankful to the National Council for Scientific and Technological Development (CNPq) for supporting this research under the Grant # 306774/2021-6, and to the Coordenação de Aperfeiçoamento de Pessoal de Nível Superior – Brasil (CAPES) – Finance Code 001.
 
 # References
