### 2024-07-12 #6620 @thiago-glissoi
Hey @danielskatz

I have made the adjustments in [commit edbde40](https://github.com/thiago-glissoi/FFF-Line-Segmentation/commit/edbde404b8d4a917dcad669eed113e5837f198f1)

> 1. Please change "cartesian" to "Cartesian".
> 2. Please add a period at the end of the Acknowledgements sentence.

I also altered the title of the zenodo deposit to match the paper title.

Thanks!
