### 2024-07-12 #6620 @danielskatz
Congratulations to @thiago-glissoi (Thiago Glissoi Lopes) and co-authors on your publication!!

JOSS relies on volunteers and we couldn't do this without you
