### 2024-07-12 #6620 @thiago-glissoi
Hey everyone 👋

I just wanted to say a big thank you to everyone who helped with this article. @bmcfee, you were an awesome editor. Even with some initial hiccups finding reviewers, you were always on top of things and super responsive.

@olivecha and @DARSakthi, thanks a ton for the review. Your feedback was spot-on, and I learned so much by working through the issues you raised.

Big shoutout to @danielskatz for your work as track editor as well.

I'm not a pro programmer, but through this submission, I learned to use Git and got the hang of GitHub. Plus, I picked up a lot of open source and general programming knowledge along the way. I'm really grateful for that.

Also, just wanted to mention that I'm a huge fan of the JOSSCast: Open Source for Researchers. Congratulations for the good work on the podcast.

Thanks again, everyone!
